chore(sdk): carve module_loader/ submodule out of runtime.rs
Pure mechanical refactor — no behavior changes, every test still
passes. Splits the 5051-line `runtime.rs` into:

- `runtime.rs` (1395 lines): Runner struct + lifecycle (new, start,
  stop, pause, resume, wait_for_signal, status, add_processor
  delegation, to_json, load_graph_file) + runtime-lifetime tests.

- `module_loader/mod.rs` (280 lines): public `impl Runner` block —
  the six public methods (load_project, load_package,
  load_workspace_packages, add_module, add_module_with,
  remove_module). Re-exports module_loader's public surface
  through `core/runtime/mod.rs` so external callers continue to
  resolve `streamlib::core::runtime::{Runner, AddModuleError,
  extract_slpkg_to_cache, host_target_triple, ...}` unchanged.

- `module_loader/strategy.rs` (334 lines): `ModuleResolverStrategy`
  enum + `resolve_strategy_to_manifest_dir` + per-tier helpers
  (lookup_installed_cache, locate_workspace_root_for_strategy,
  workspace_stage_dir, read_version_from_manifest_dir,
  read_module_ident_from_manifest_dir, read_module_ident_from_slpkg,
  check_cdylib_present_when_rust_impl,
  split_canonical_for_legacy_error).

- `module_loader/recursive_walker.rs` (250 lines): the recursive
  dep walker with cycle detection — `add_module_recursively`,
  `add_module_recursive_body`, and `derive_dep_strategy_and_ident`.
  All three are free functions taking `&Iceoryx2Node` instead of
  `&Runner`, so the recursion doesn't need to thread `&self`.

- `module_loader/processor_registration.rs` (439 lines): the
  `register_manifest_processors` function (Python venv prep, Rust
  cdylib dlopen, PROCESSOR_REGISTRY.register_dynamic) + the
  process-lifetime `LOADED_PLUGIN_LIBRARIES` static + the public
  `host_target_triple` + the `list_available_triples` helper.

- `module_loader/schema_registration.rs` (190 lines):
  `register_package_schemas` + `resolve_config_schema_canonical_id`
  + private helpers (canonical_identifier_for_schema,
  strip_legacy_semver_suffix).

- `module_loader/slpkg.rs` (98 lines): the public
  `extract_slpkg_to_cache` archive extractor.

- `module_loader/workspace.rs` (71 lines): `CanonicalPackageId` +
  `parse_canonical_package_id` + `resolve_workspace_root`.

- `module_loader/errors.rs` (324 lines): all three typed error
  enums (`AddModuleError`, `RemoveModuleError`,
  `LoadWorkspacePackagesError`) + their `From<E> for Error` impls.

- `module_loader/tests.rs` (1549 lines): every module-loading test
  — strategy conformance, recursive walker, cycle detection,
  parse_canonical_package_id, resolve_workspace_root,
  list_available_triples_filters_to_subdirs_and_sorts,
  load_workspace_packages negative paths, the full add_module_tests
  module with all 19 add_module / add_module_with cases. Lifecycle
  tests stay in runtime.rs's tests module.

Motivation: runtime.rs reached 5051 lines after #1048's additions
and was running into agent-editability ceilings (Read tool's
25k-token cap, Edit-tool's uniqueness requirement getting harder
at scale). Per CLAUDE.md "no bad patterns left behind on engine
changes": establishing the new ModuleResolverStrategy primitive
in an unreadable file would leave a bad pattern for downstream
PRs to inherit.

Visibility: every type that was public before stays public (via
re-export through `core/runtime/mod.rs`). Private helpers move
to `pub(super)` so tests in `module_loader/tests.rs` can reach
them. No external callers need to change.

Verified:
- `cargo test -p streamlib-engine --lib 'core::runtime'` — 52/52 pass
- `cargo test -p streamlib-engine --test pack_then_load_smoke` — 4/4 pass
- `cargo check --workspace --all-targets` — clean
- `cargo xtask check-boundaries` — 4373 files, no violations
- `cargo xtask lint-logging` — 462 files, no violations
