+++ libs/streamlib-engine/src/core/runtime/mod.rs
@@ -2,6 +2,7 @@
 // SPDX-License-Identifier: BUSL-1.1
 
 mod graph_change_listener;
+mod module_loader;
 mod operations;
 mod operations_runtime;
 #[allow(clippy::module_inception)]
@@ -9,10 +10,11 @@ mod runtime;
 mod runtime_unique_id;
 mod status;
 
-pub use operations::{BoxFuture, RuntimeOperations};
-pub use runtime::{
+pub use module_loader::{
     extract_slpkg_to_cache, host_target_triple, AddModuleError, LoadWorkspacePackagesError,
-    RemoveModuleError, Runner,
+    ModuleResolverStrategy, RemoveModuleError,
 };
+pub use operations::{BoxFuture, RuntimeOperations};
+pub use runtime::Runner;
 pub use runtime_unique_id::RuntimeUniqueId;
 pub use status::RuntimeStatus;
