+++ libs/streamlib-engine/src/core/runtime/runtime.rs
@@ -1,7 +1,6 @@
 // Copyright (c) 2025 Jonathan Fontanez
 // SPDX-License-Identifier: BUSL-1.1
 
-use std::collections::HashSet;
 use std::ops::ControlFlow;
 use std::sync::atomic::{AtomicBool, Ordering};
 use std::sync::Arc;
@@ -28,13 +27,6 @@ use crate::core::processors::ProcessorState;
 use crate::core::pubsub::{topics, Event, EventListener, ProcessorEvent, RuntimeEvent, PUBSUB};
 use crate::core::{InputLinkPortRef, OutputLinkPortRef, Result, Error};
 use crate::iceoryx2::Iceoryx2Node;
-
-/// Keeps loaded dylib plugin libraries alive for the process lifetime.
-///
-/// When a Rust dylib plugin is loaded via `load_project()`, the `Library` handle
-/// must remain alive so that the registered processor vtables stay valid.
-static LOADED_PLUGIN_LIBRARIES: std::sync::LazyLock<parking_lot::Mutex<Vec<libloading::Library>>> =
-    std::sync::LazyLock::new(|| parking_lot::Mutex::new(Vec::new()));
 
 /// Storage variant for tokio runtime in Runner.
 ///
@@ -348,840 +340,8 @@ impl Runner {
     }
 
     // =========================================================================
-    // Module Loading — public API surface
+    // Module Loading — see `core/runtime/module_loader/` for the impl.
     // =========================================================================
-    //
-    // Every module-loading entry point on `Runner` routes through
-    // `add_module_with(ident, strategy)`. The three deprecated methods
-    // (`load_project` / `load_package` / `load_workspace_packages`) stay
-    // functional as one-line wrappers that construct the matching
-    // `ModuleResolverStrategy` variant and dispatch through the unified
-    // resolver. They will move behind a non-public visibility (or be
-    // deleted) in a later cleanup; the body lives in the strategy
-    // resolver below.
-
-    /// Extract a `.slpkg` archive, then load the manifest it contains.
-    ///
-    /// One-line wrapper around
-    /// [`Self::add_module_with`] with
-    /// [`ModuleResolverStrategy::SlpkgArchive`]. The strategy's resolver
-    /// extracts the archive to the package cache, reads the embedded
-    /// manifest, then drives the unified module-load flow.
-    pub fn load_package(&self, slpkg_path: impl AsRef<std::path::Path>) -> Result<()> {
-        let slpkg_path = slpkg_path.as_ref().to_path_buf();
-        // Pre-read the manifest so we can build the strict
-        // `ModuleIdent` the unified flow requires. The `.slpkg`'s
-        // embedded `streamlib.yaml` is authoritative for the identity
-        // — `add_module_with` then re-verifies that the extracted
-        // manifest matches what we declared here, so the round-trip is
-        // self-checking.
-        let ident = read_module_ident_from_slpkg(&slpkg_path)?;
-        self.add_module_with(
-            ident,
-            ModuleResolverStrategy::SlpkgArchive { path: slpkg_path },
-        )?;
-        Ok(())
-    }
-
-    /// Load the manifest at a directory containing `streamlib.yaml`.
-    ///
-    /// One-line wrapper around
-    /// [`Self::add_module_with`] with
-    /// [`ModuleResolverStrategy::ManifestDirectory`]. The strategy's
-    /// resolver returns the directory as-is; the manifest's own
-    /// `[package]` block supplies the identity the unified flow
-    /// validates against.
-    pub fn load_project(&self, project_path: impl AsRef<std::path::Path>) -> Result<()> {
-        let project_path = project_path.as_ref().to_path_buf();
-        let ident = read_module_ident_from_manifest_dir(&project_path)?;
-        self.add_module_with(
-            ident,
-            ModuleResolverStrategy::ManifestDirectory { path: project_path },
-        )?;
-        Ok(())
-    }
-
-    /// Look up workspace-staged packages by canonical id and route each
-    /// through [`Self::add_module_with`] with
-    /// [`ModuleResolverStrategy::WorkspaceStaged`].
-    ///
-    /// `cargo xtask build-plugins` must have run first — the strategy's
-    /// resolver errors with [`LoadWorkspacePackagesError::PackageNotStaged`]
-    /// when a name's staged dir is missing.
-    ///
-    /// Workspace root resolution: `STREAMLIB_WORKSPACE_ROOT` env var
-    /// when set (and the path exists), otherwise
-    /// `cargo locate-project --workspace`.
-    pub fn load_workspace_packages<I, S>(
-        &self,
-        names: I,
-    ) -> std::result::Result<(), LoadWorkspacePackagesError>
-    where
-        I: IntoIterator<Item = S>,
-        S: AsRef<str>,
-    {
-        // Eagerly validate every id BEFORE touching the workspace root —
-        // a typo'd id should surface as `InvalidPackageId` immediately
-        // rather than masquerading as `WorkspaceRootNotFound` when the
-        // env var is unset.
-        let parsed: Vec<(String, streamlib_idents::PackageRef)> = names
-            .into_iter()
-            .map(|name| {
-                let name_str = name.as_ref().to_string();
-                let parsed = parse_canonical_package_id(&name_str)?;
-                let pkg_ref = streamlib_idents::PackageRef::new(
-                    streamlib_idents::Org::new(parsed.org_str).map_err(|_| {
-                        LoadWorkspacePackagesError::InvalidPackageId(name_str.clone())
-                    })?,
-                    streamlib_idents::Package::new(parsed.name_str).map_err(|_| {
-                        LoadWorkspacePackagesError::InvalidPackageId(name_str.clone())
-                    })?,
-                );
-                Ok::<_, LoadWorkspacePackagesError>((name_str, pkg_ref))
-            })
-            .collect::<std::result::Result<_, _>>()?;
-
-        // Surface workspace-root failure once, up front, so the error
-        // mode matches the historical helper's contract (single
-        // `WorkspaceRootNotFound` rather than one-per-id).
-        let _ = resolve_workspace_root()?;
-
-        for (name_str, pkg_ref) in parsed {
-            // Keep the typed segments around for the back-compat error
-            // translation below — moving them into ModuleIdent::any
-            // would consume the values one statement before we need
-            // them for the LegacyPackageIdentityMismatch arm.
-            let req_org = pkg_ref.org.as_str().to_string();
-            let req_name = pkg_ref.name.as_str().to_string();
-            let ident = streamlib_idents::ModuleIdent::any(pkg_ref.org, pkg_ref.name);
-            self.add_module_with(ident, ModuleResolverStrategy::WorkspaceStaged)
-                .map_err(|err| match err {
-                    AddModuleError::WorkspaceStageMiss { expected_path, .. } => {
-                        LoadWorkspacePackagesError::PackageNotStaged {
-                            name: name_str.clone(),
-                            expected_path,
-                        }
-                    }
-                    AddModuleError::ManifestIdentityMismatch {
-                        source_path, actual, ..
-                    } => {
-                        let (actual_org, actual_name) = split_canonical_for_legacy_error(&actual);
-                        LoadWorkspacePackagesError::PackageIdentityMismatch {
-                            staged_path: source_path,
-                            requested_org: req_org.clone(),
-                            requested_name: req_name.clone(),
-                            actual_org,
-                            actual_name,
-                        }
-                    }
-                    AddModuleError::CdylibMissingForRustImpl { expected_path, .. } => {
-                        LoadWorkspacePackagesError::CdylibMissing {
-                            name: name_str.clone(),
-                            expected_path,
-                        }
-                    }
-                    AddModuleError::LoadProjectFailed { source, .. } => {
-                        LoadWorkspacePackagesError::LoadProjectFailed {
-                            name: name_str.clone(),
-                            source,
-                        }
-                    }
-                    AddModuleError::WorkspaceRootInvalid { .. } => {
-                        LoadWorkspacePackagesError::WorkspaceRootNotFound
-                    }
-                    other => LoadWorkspacePackagesError::LoadProjectFailed {
-                        name: name_str.clone(),
-                        source: Box::new(other.into()),
-                    },
-                })?;
-        }
-
-        Ok(())
-    }
-
-    /// Load a `streamlib.yaml`-packaged module by typed
-    /// [`streamlib_idents::ModuleIdent`]. Routes through the default
-    /// resolver chain (workspace stage → installed-package cache).
-    ///
-    /// Imperative complement to the yaml-driven path: both this and
-    /// [`Self::load_project`] drive into the same internal
-    /// module-loading machinery; the yaml form is for declarative
-    /// deployment manifests, the imperative form is for REST endpoints,
-    /// hot-reload tools, test setup, and composition-library wrapping.
-    ///
-    /// Transitive dependencies declared in the loaded module's
-    /// `streamlib.yaml` are recursively routed through the same flow,
-    /// each picking the dep's appropriate strategy (declared `path:` →
-    /// [`ModuleResolverStrategy::ManifestDirectory`]; consumer `patch:`
-    /// override of any kind likewise; bare registry/git declarations →
-    /// [`ModuleResolverStrategy::DefaultChain`]). Cycles are detected
-    /// per-call and surfaced as
-    /// [`AddModuleError::DependencyCycleDetected`].
-    ///
-    /// Calls are idempotent at the registry layer: re-loading a module
-    /// whose processors / schemas are already registered surfaces no
-    /// error and re-runs the dylib's plugin callback (which the engine
-    /// already tolerates per `register_dynamic`'s dedup semantics).
-    pub fn add_module(
-        &self,
-        module: streamlib_idents::ModuleIdent,
-    ) -> std::result::Result<(), AddModuleError> {
-        self.add_module_with(module, ModuleResolverStrategy::DefaultChain)
-    }
-
-    /// Load a module via an explicit [`ModuleResolverStrategy`].
-    ///
-    /// In-code equivalent of the manifest's `patch:` table — callers
-    /// can pin a dep to a workspace path, a `.slpkg` archive, or a
-    /// specific resolver tier without editing yaml. See
-    /// [`ModuleResolverStrategy`] for the variant catalog.
-    #[tracing::instrument(name = "runtime.add_module_with", skip(self), fields(module = %module, strategy = ?strategy))]
-    pub fn add_module_with(
-        &self,
-        module: streamlib_idents::ModuleIdent,
-        strategy: ModuleResolverStrategy,
-    ) -> std::result::Result<(), AddModuleError> {
-        // Two collections threaded through the recursion:
-        // - `seen` is O(1) cycle membership lookup
-        // - `path` preserves insertion order so cycle errors carry the
-        //   full recursion edge (`A → B → A` rather than just `A`)
-        let mut seen: HashSet<streamlib_idents::PackageRef> = HashSet::new();
-        let mut path: Vec<streamlib_idents::PackageRef> = Vec::new();
-        self.add_module_recursively(module, strategy, &mut seen, &mut path)
-    }
-
-    /// Recursive worker: resolves the strategy, validates the
-    /// manifest's identity + version range, registers the package's
-    /// schemas, walks dependencies (each routed through this same
-    /// helper), then registers the package's processors.
-    ///
-    /// `seen` tracks every [`PackageRef`] currently on the recursion
-    /// stack — a hit means a dep cycle and surfaces
-    /// [`AddModuleError::DependencyCycleDetected`]. The current
-    /// package's ref is inserted on entry and removed on exit, so
-    /// sibling sub-trees can revisit a shared transitive dep without
-    /// false positives.
-    ///
-    /// Order matters: deps load before this package's own processors so
-    /// schemas referenced from `with_config_schema(...)` resolve.
-    fn add_module_recursively(
-        &self,
-        module: streamlib_idents::ModuleIdent,
-        strategy: ModuleResolverStrategy,
-        seen: &mut HashSet<streamlib_idents::PackageRef>,
-        path: &mut Vec<streamlib_idents::PackageRef>,
-    ) -> std::result::Result<(), AddModuleError> {
-        let pkg_ref = module.package_ref();
-        if !seen.insert(pkg_ref.clone()) {
-            // Reaching this package while it is already mid-load on
-            // the recursion stack is the dependency-cycle signal.
-            // Surface the full recursion path plus the repeated vertex
-            // so the caller can see exactly which edge re-entered.
-            let mut cycle = path.clone();
-            cycle.push(pkg_ref);
-            return Err(AddModuleError::DependencyCycleDetected { cycle });
-        }
-        path.push(pkg_ref.clone());
-        // Run the body, then remove `pkg_ref` from both `seen` and
-        // `path` regardless of the body's exit path — sibling
-        // sub-trees can revisit a shared transitive dep without
-        // false-positive cycle reports.
-        let result = self.add_module_recursive_body(module, strategy, seen, path);
-        seen.remove(&pkg_ref);
-        path.pop();
-        result
-    }
-
-    /// Body of [`Self::add_module_recursively`] — split out so the
-    /// caller can pop `pkg_ref` from `seen` + `path` after every
-    /// exit path.
-    fn add_module_recursive_body(
-        &self,
-        module: streamlib_idents::ModuleIdent,
-        strategy: ModuleResolverStrategy,
-        seen: &mut HashSet<streamlib_idents::PackageRef>,
-        path: &mut Vec<streamlib_idents::PackageRef>,
-    ) -> std::result::Result<(), AddModuleError> {
-        use crate::core::config::ProjectConfig;
-
-        let pkg_ref = module.package_ref();
-        let (manifest_dir, on_disk_version) =
-            resolve_strategy_to_manifest_dir(&strategy, Some(&pkg_ref))?;
-
-        // Read the manifest; this is the authoritative source of
-        // identity for the package at the resolved location.
-        let config = ProjectConfig::load(&manifest_dir)
-            .map_err(|e| AddModuleError::ManifestLoadFailed {
-                module: module.clone(),
-                source_path: manifest_dir.clone(),
-                detail: e.to_string(),
-            })?;
-
-        config
-            .check_streamlib_version_compatibility()
-            .map_err(|e| AddModuleError::ManifestLoadFailed {
-                module: module.clone(),
-                source_path: manifest_dir.clone(),
-                detail: e.to_string(),
-            })?;
-
-        // Identity check: the manifest's `[package]` org/name must
-        // match the requested ident. Catches `.slpkg`s shipped with
-        // the wrong content as well as workspace-stage clobbers.
-        let pkg_meta = config
-            .package
-            .as_ref()
-            .ok_or_else(|| AddModuleError::ManifestLoadFailed {
-                module: module.clone(),
-                source_path: manifest_dir.clone(),
-                detail: "manifest has no `package:` block".into(),
-            })?;
-        if pkg_meta.org != module.org || pkg_meta.name != module.name {
-            return Err(AddModuleError::ManifestIdentityMismatch {
-                module: module.clone(),
-                source_path: manifest_dir.clone(),
-                actual: format!(
-                    "@{}/{}",
-                    pkg_meta.org.as_str(),
-                    pkg_meta.name.as_str()
-                ),
-            });
-        }
-        if !module.version.matches(on_disk_version) {
-            return Err(AddModuleError::VersionRangeUnsatisfied {
-                module: module.clone(),
-                found: on_disk_version,
-                source_path: manifest_dir.clone(),
-            });
-        }
-
-        tracing::info!(
-            "add_module: '{}' → {} (on-disk version {})",
-            module,
-            manifest_dir.display(),
-            on_disk_version,
-        );
-
-        // Register every schema declared in this package's
-        // `streamlib.yaml`'s `schemas:` list BEFORE recursing — schemas
-        // are leaves in the dep graph and don't depend on transitive
-        // package state. (Processor registration runs AFTER deps so
-        // bare-name config schemas in dep packages have already
-        // populated the registry.)
-        register_package_schemas(&manifest_dir, &config).map_err(|e| {
-            AddModuleError::LoadProjectFailed {
-                module: module.clone(),
-                source: Box::new(e),
-            }
-        })?;
-
-        // Walk transitive deps. Each dep is itself routed through
-        // `add_module_recursively`, so the per-dep strategy lookup is
-        // the same shape as the consumer's own top-level call.
-        for (dep_ref, spec) in &config.dependencies {
-            let (dep_ident, dep_strategy) = self
-                .derive_dep_strategy_and_ident(&manifest_dir, dep_ref, spec, &config.patch)
-                .map_err(|e| AddModuleError::LoadProjectFailed {
-                    module: module.clone(),
-                    source: Box::new(e),
-                })?;
-            tracing::info!(
-                "Loading dependency '{}' (strategy {:?})",
-                dep_ident,
-                dep_strategy
-            );
-            self.add_module_recursively(dep_ident, dep_strategy, seen, path)?;
-        }
-
-        // Now register this package's own processors.
-        self.register_manifest_processors(&manifest_dir, &config)
-            .map_err(|e| AddModuleError::LoadProjectFailed {
-                module: module.clone(),
-                source: Box::new(e),
-            })?;
-
-        Ok(())
-    }
-
-    /// Map a single declared dep (with optional consumer `patch:`
-    /// override) to the [`ModuleIdent`] + [`ModuleResolverStrategy`]
-    /// pair that recursively re-enters [`Self::add_module_recursively`].
-    ///
-    /// Patch precedence mirrors Cargo's `[patch.crates-io]`:
-    /// consumer's patch table wins when present; otherwise the dep
-    /// declaration's source variant decides.
-    fn derive_dep_strategy_and_ident(
-        &self,
-        consumer_dir: &std::path::Path,
-        dep_ref: &streamlib_idents::PackageRef,
-        spec: &streamlib_idents::DependencySpec,
-        patch: &std::collections::BTreeMap<
-            streamlib_idents::PackageRef,
-            streamlib_idents::DependencySpec,
-        >,
-    ) -> Result<(streamlib_idents::ModuleIdent, ModuleResolverStrategy)> {
-        use streamlib_idents::{DependencySpec, ModuleIdent, SemVerRange};
-
-        // The declared range (registry deps) constrains version
-        // resolution even when the source is patched. Path / git deps
-        // don't carry a range at declaration time — the patched
-        // location's manifest version becomes authoritative
-        // (range = any).
-        let declared_range = match spec {
-            DependencySpec::Registry(r) => r.version.clone(),
-            DependencySpec::Path(_) | DependencySpec::Git(_) => SemVerRange::Any,
-        };
-
-        let (strategy_spec, source_label) = match patch.get(dep_ref) {
-            Some(patched) => (patched, "patch"),
-            None => (spec, "dep"),
-        };
-
-        let strategy = match strategy_spec {
-            DependencySpec::Path(p) => {
-                let abs = if p.path.is_absolute() {
-                    p.path.clone()
-                } else {
-                    consumer_dir.join(&p.path)
-                };
-                // npm/wrangler-style strict validation: a missing path
-                // is a hard error so the dev knows immediately to fix
-                // the manifest.
-                if source_label == "patch" && !abs.exists() {
-                    return Err(Error::Configuration(format!(
-                        "patch entry for '{dep_ref}' in {}/{} points at `{}` \
-                         which does not exist. Path patches are dev-time \
-                         overrides — they must resolve to a real directory \
-                         at parse time.",
-                        consumer_dir.display(),
-                        crate::core::config::ProjectConfig::FILE_NAME,
-                        abs.display(),
-                    )));
-                }
-                ModuleResolverStrategy::ManifestDirectory { path: abs }
-            }
-            DependencySpec::Git(g) => {
-                let cache_dir =
-                    crate::core::streamlib_home::get_streamlib_home().join("resolver-cache");
-                let checkout = streamlib_idents::fetch_git(
-                    &dep_ref.to_string(),
-                    &g.git,
-                    &g.rev,
-                    &cache_dir,
-                )
-                .map_err(|e| Error::Configuration(e.to_string()))?;
-                ModuleResolverStrategy::ManifestDirectory { path: checkout }
-            }
-            DependencySpec::Registry(_) => {
-                if source_label == "patch" {
-                    return Err(Error::Configuration(format!(
-                        "patch entry for '{dep_ref}' in {}/{} is registry-flavored. \
-                         The v1 resolver doesn't ship a registry — declare a \
-                         `path:` or `git:` patch entry, or remove the patch and \
-                         rely on the installed-package cache.",
-                        consumer_dir.display(),
-                        crate::core::config::ProjectConfig::FILE_NAME,
-                    )));
-                }
-                ModuleResolverStrategy::DefaultChain
-            }
-        };
-
-        let ident = ModuleIdent::new(dep_ref.org.clone(), dep_ref.name.clone(), declared_range);
-        Ok((ident, strategy))
-    }
-
-    /// Register this package's processors with the global processor
-    /// registry. Non-recursive — the caller is responsible for having
-    /// walked any transitive deps first.
-    fn register_manifest_processors(
-        &self,
-        project_path: &std::path::Path,
-        config: &crate::core::config::ProjectConfig,
-    ) -> Result<()> {
-        use crate::core::compiler::compiler_ops::create_deno_subprocess_host_constructor;
-        use crate::core::compiler::compiler_ops::create_python_native_subprocess_host_constructor;
-        use crate::core::compiler::compiler_ops::ensure_processor_venv;
-        use crate::core::compiler::compiler_ops::resolve_python_native_lib_path;
-        use crate::core::config::ProjectConfig;
-        use crate::core::descriptors::{PortDescriptor, ProcessorRuntime};
-        use crate::core::execution::{ExecutionConfig, ProcessExecution};
-        use crate::core::ProcessorDescriptor;
-
-        if config.processors.is_empty() {
-            tracing::debug!(
-                "No processors declared in {} at {} (schemas-only package or fixture).",
-                ProjectConfig::FILE_NAME,
-                project_path.display()
-            );
-            return Ok(());
-        }
-
-        // Eagerly create venv for Python packages so processors don't race at spawn time
-        let has_python_processors = config.processors.iter().any(|p| {
-            matches!(
-                p.runtime.language,
-                streamlib_processor_schema::ProcessorLanguage::Python
-            )
-        });
-
-        if has_python_processors {
-            let package_label = config
-                .package
-                .as_ref()
-                .map(|p| p.name.as_str())
-                .unwrap_or("unknown");
-            tracing::info!("Pre-creating Python venv for package '{}'", package_label);
-            ensure_processor_venv(package_label, project_path)?;
-        }
-
-        let mut registered_count = 0usize;
-        let mut rust_dylib_loaded = false;
-
-        // Resolve package metadata once per project — every dynamically
-        // registered processor in this manifest uses these typed segments
-        // to compose its full structured `SchemaIdent`.
-        let package_metadata = config.package.as_ref().ok_or_else(|| {
-            Error::Configuration(format!(
-                "{} at {} declares processors but is missing a `package:` block. \
-                 The structured-everywhere processor identity rule requires \
-                 `package: {{ org, name, version }}` to compose each processor's \
-                 SchemaIdent.",
-                ProjectConfig::FILE_NAME,
-                project_path.display(),
-            ))
-        })?;
-
-        // Eagerly resolve the dependency graph once per project — used
-        // below by `resolve_config_schema_canonical_id` for each
-        // processor's bare-name config schema (#767). The resolver walks
-        // declared paths/git/.slpkg sources and validates the `schemas:`
-        // map; we only invoke it when at least one processor actually
-        // declares a config block.
-        let config_resolved: Option<streamlib_idents::ResolvedPackages> =
-            if config.processors.iter().any(|p| p.config.is_some()) {
-                Some(
-                    streamlib_idents::resolve_with(
-                        project_path,
-                        &streamlib_idents::ResolverOptions::default(),
-                    )
-                    .map_err(|e| {
-                        Error::Configuration(format!(
-                            "failed to resolve manifest dependencies for bare-name \
-                             config schema lookup at {}: {}",
-                            project_path.display(),
-                            e
-                        ))
-                    })?,
-                )
-            } else {
-                None
-            };
-
-        for proc_schema in &config.processors {
-            // Compose the structured processor ident from the manifest's
-            // package metadata + the processor's PascalCase short name.
-            let proc_type_name = streamlib_processor_schema::TypeName::new(&proc_schema.name)
-                .map_err(|e| {
-                    Error::Configuration(format!(
-                        "processor short name `{}` in {} is not valid PascalCase: {}",
-                        proc_schema.name,
-                        project_path.display(),
-                        e
-                    ))
-                })?;
-            let proc_schema_ident = crate::core::descriptors::SchemaIdent::new(
-                package_metadata.org.clone(),
-                package_metadata.name.clone(),
-                proc_type_name,
-                package_metadata.version.clone(),
-            );
-
-            // Map runtime language to ProcessorRuntime
-            let runtime = match proc_schema.runtime.language {
-                streamlib_processor_schema::ProcessorLanguage::Python => ProcessorRuntime::Python,
-                streamlib_processor_schema::ProcessorLanguage::TypeScript => {
-                    ProcessorRuntime::TypeScript
-                }
-                streamlib_processor_schema::ProcessorLanguage::Rust => {
-                    // Rust dylib plugins self-register via export_plugin! macro.
-                    // Load the dylib once per project (all Rust processors in the
-                    // same YAML share one dylib), then validate each processor
-                    // was actually registered.
-                    if !rust_dylib_loaded {
-                        let host_triple = host_target_triple();
-                        let lib_dir = project_path.join("lib");
-                        let triple_dir = lib_dir.join(host_triple);
-                        let dylib_ext = if cfg!(target_os = "macos") {
-                            "dylib"
-                        } else if cfg!(target_os = "windows") {
-                            "dll"
-                        } else {
-                            "so"
-                        };
-
-                        let dylib_path = std::fs::read_dir(&triple_dir)
-                            .map_err(|e| {
-                                // If `lib/` exists but the triple-keyed
-                                // subdir is absent, surface the available
-                                // triples so the user sees exactly which
-                                // platforms this slpkg was packed for.
-                                let available =
-                                    list_available_triples(&lib_dir).unwrap_or_default();
-                                let detail = if available.is_empty() {
-                                    format!(
-                                        "Failed to read {}: {}. \
-                                         The package may be missing a Rust artifact for this host.",
-                                        triple_dir.display(),
-                                        e
-                                    )
-                                } else {
-                                    format!(
-                                        "Failed to read {}: {}. \
-                                         This slpkg was packed for: [{}]. \
-                                         The host triple is `{}` — repack on a matching host \
-                                         or run this pipeline on a host that matches one of the \
-                                         packed triples.",
-                                        triple_dir.display(),
-                                        e,
-                                        available.join(", "),
-                                        host_triple
-                                    )
-                                };
-                                Error::Configuration(detail)
-                            })?
-                            .filter_map(|entry| entry.ok())
-                            .map(|entry| entry.path())
-                            .find(|path| path.extension().is_some_and(|ext| ext == dylib_ext))
-                            .ok_or_else(|| {
-                                Error::Configuration(format!(
-                                    "No .{} file found in {} (host triple `{}`)",
-                                    dylib_ext,
-                                    triple_dir.display(),
-                                    host_triple
-                                ))
-                            })?;
-
-                        tracing::info!("Loading Rust dylib plugin: {}", dylib_path.display());
-
-                        // Safety: Loading a dynamic library is inherently unsafe.
-                        // The dylib must be a valid StreamLib plugin built with
-                        // a compatible streamlib-plugin-abi version.
-                        let lib = unsafe {
-                            libloading::Library::new(&dylib_path).map_err(|e| {
-                                Error::Configuration(format!(
-                                    "Failed to load dylib {}: {}",
-                                    dylib_path.display(),
-                                    e
-                                ))
-                            })?
-                        };
-
-                        let decl: &streamlib_plugin_abi::PluginDeclaration = unsafe {
-                            let symbol = lib
-                                .get::<*const streamlib_plugin_abi::PluginDeclaration>(
-                                    b"STREAMLIB_PLUGIN\0",
-                                )
-                                .map_err(|e| {
-                                    Error::Configuration(format!(
-                                        "Plugin '{}' missing STREAMLIB_PLUGIN symbol. \
-                                         Ensure the plugin uses the export_plugin! macro: {}",
-                                        dylib_path.display(),
-                                        e
-                                    ))
-                                })?;
-                            &**symbol
-                        };
-
-                        if decl.abi_version != streamlib_plugin_abi::STREAMLIB_ABI_VERSION {
-                            return Err(Error::Configuration(format!(
-                                "ABI version mismatch for '{}': plugin has v{}, \
-                                 runtime expects v{}. Rebuild the plugin with a \
-                                 compatible streamlib-plugin-abi version.",
-                                dylib_path.display(),
-                                decl.abi_version,
-                                streamlib_plugin_abi::STREAMLIB_ABI_VERSION,
-                            )));
-                        }
-
-                        // Build the HostServices payload from the host's
-                        // process-wide statics + this runtime's iceoryx2
-                        // node, hand it to the cdylib's register callback.
-                        // The cdylib's macro-emitted prologue calls
-                        // `install_host_services` which bridges every
-                        // per-DSO static (tracing dispatch, PUBSUB,
-                        // schema registry, iceoryx2 logger) into the
-                        // host's instances before processor registration.
-                        let host_services =
-                            crate::core::plugin::host_services::runtime_facing::host_services_for_self(
-                                &self.iceoryx2_node,
-                            );
-                        // SAFETY: `host_services` outlives the call;
-                        // the cdylib's callback returns before this
-                        // function frame is dropped.
-                        unsafe {
-                            (decl.register)(
-                                &host_services as *const _ as *const ::std::ffi::c_void,
-                            );
-                        }
-
-                        // Keep the library alive for the process lifetime
-                        LOADED_PLUGIN_LIBRARIES.lock().push(lib);
-
-                        rust_dylib_loaded = true;
-                        tracing::info!(
-                            "Rust dylib plugin loaded and registered: {}",
-                            dylib_path.display()
-                        );
-                    }
-
-                    // Validate the processor was registered by the dylib —
-                    // compare by structured `SchemaIdent`, not bare PascalCase
-                    // (two packages can declare the same short name).
-                    let registered = crate::core::processors::PROCESSOR_REGISTRY
-                        .list_registered()
-                        .iter()
-                        .any(|desc| desc.name == proc_schema_ident);
-                    if !registered {
-                        return Err(Error::Configuration(format!(
-                            "Processor '{}' declared in streamlib.yaml but not \
-                             registered by the dylib. Ensure export_plugin!() \
-                             includes this processor.",
-                            proc_schema_ident
-                        )));
-                    }
-
-                    tracing::info!("Validated Rust dylib processor '{}'", proc_schema.name);
-                    registered_count += 1;
-                    continue;
-                }
-            };
-
-            let inputs: Vec<PortDescriptor> = proc_schema
-                .inputs
-                .iter()
-                .map(|p| {
-                    PortDescriptor::new(
-                        &p.name,
-                        p.description.as_deref().unwrap_or(""),
-                        p.schema.clone(),
-                        true,
-                    )
-                })
-                .collect();
-
-            let outputs: Vec<PortDescriptor> = proc_schema
-                .outputs
-                .iter()
-                .map(|p| {
-                    PortDescriptor::new(
-                        &p.name,
-                        p.description.as_deref().unwrap_or(""),
-                        p.schema.clone(),
-                        true,
-                    )
-                })
-                .collect();
-
-            let mut descriptor = ProcessorDescriptor::new(
-                proc_schema_ident.clone(),
-                proc_schema.description.as_deref().unwrap_or(""),
-            )
-            .with_version(&proc_schema.version)
-            .with_runtime(runtime.clone());
-
-            if let Some(entrypoint) = &proc_schema.entrypoint {
-                descriptor = descriptor.with_entrypoint(entrypoint);
-            }
-
-            if let Some(config) = &proc_schema.config {
-                // Resolve the bare-name `TypeName` (#767) against the
-                // manifest's `schemas:` map to its canonical id string.
-                // The lookup walks the manifest's declarations, locates
-                // the owning package + schema file, and reads the
-                // schema's `metadata.type` (new shape) or
-                // `metadata.name` (legacy reverse-DNS) to compose the
-                // id. This must match what
-                // `register_package_schemas`/`canonical_identifier_for_schema`
-                // registered when the same manifest was loaded.
-                let canonical =
-                    resolve_config_schema_canonical_id(project_path, config, &config_resolved)
-                        .map_err(|msg| {
-                            Error::Configuration(format!(
-                                "processor `{}` config schema `{}`: {}",
-                                proc_schema.name,
-                                config.schema.as_str(),
-                                msg
-                            ))
-                        })?;
-                descriptor = descriptor.with_config_schema(canonical);
-            }
-
-            if let Some(scheduling) = &proc_schema.scheduling {
-                descriptor = descriptor.with_scheduling(scheduling.clone());
-            }
-
-            descriptor.inputs = inputs;
-            descriptor.outputs = outputs;
-
-            // Convert schema execution mode to runtime ExecutionConfig
-            let execution = match &proc_schema.execution {
-                streamlib_processor_schema::ProcessorSchemaExecution::Reactive => {
-                    ProcessExecution::Reactive
-                }
-                streamlib_processor_schema::ProcessorSchemaExecution::Manual => {
-                    ProcessExecution::Manual
-                }
-                streamlib_processor_schema::ProcessorSchemaExecution::Continuous { interval_ms } => {
-                    ProcessExecution::Continuous {
-                        interval_ms: *interval_ms,
-                    }
-                }
-            };
-            let execution_config = ExecutionConfig::new(execution);
-
-            // Create constructor based on runtime language.
-            // Python and TypeScript subprocesses both use native FFI for direct
-            // iceoryx2 access — no pipe bridge for data I/O.
-            let constructor = match runtime {
-                ProcessorRuntime::Python => {
-                    let native_lib_path = resolve_python_native_lib_path()?;
-                    create_python_native_subprocess_host_constructor(
-                        &descriptor,
-                        execution_config,
-                        project_path.to_path_buf(),
-                        native_lib_path,
-                    )
-                }
-                ProcessorRuntime::TypeScript => create_deno_subprocess_host_constructor(
-                    &descriptor,
-                    execution_config,
-                    project_path.to_path_buf(),
-                ),
-                _ => unreachable!(),
-            };
-
-            crate::core::processors::PROCESSOR_REGISTRY
-                .register_dynamic(descriptor, constructor)?;
-
-            tracing::info!(
-                "Registered processor '{}' ({:?})",
-                proc_schema.name,
-                runtime
-            );
-
-            registered_count += 1;
-        }
-
-        tracing::info!(
-            "Loaded {} processor(s) from {}",
-            registered_count,
-            project_path.display()
-        );
-
-        Ok(())
-    }
-
     // =========================================================================
     // Lifecycle
     // =========================================================================
@@ -1840,345 +1000,8 @@ impl Runner {
         self.load_graph_file(&def)
     }
 
-    /// Unload a previously-added module.
-    ///
-    /// **Not yet implemented.** Module-level unload requires the
-    /// hot-reload lifecycle work that's explicitly out of scope for
-    /// the current All-Dynamic Package Loading milestone — see the
-    /// milestone's "Explicitly out of scope (deferred to later
-    /// milestones)" section ("`unload_package` / hot-reload
-    /// lifecycle. Load-only, runtime-lifetime registration this
-    /// milestone."). The method exists as an explicit boundary
-    /// marker (rather than being absent) so AI agents and other
-    /// callers reaching for it from the `add_module` counterpart get
-    /// a typed error pointing at the milestone gap instead of a
-    /// `method not found` diagnostic.
-    ///
-    /// Calling this returns
-    /// [`RemoveModuleError::HotReloadLifecycleNotYetImplemented`]
-    /// without altering any runtime state.
-    pub fn remove_module(
-        &self,
-        module: streamlib_idents::ModuleIdent,
-    ) -> std::result::Result<(), RemoveModuleError> {
-        Err(RemoveModuleError::HotReloadLifecycleNotYetImplemented { module })
-    }
 }
 
-/// Iterate `config.schemas` map entries, registering each `Local` schema
-/// (the YAML body keyed by its canonical identifier) with the engine's
-/// runtime schema registry. `External` entries are import declarations
-/// owned by other packages and are skipped here — the dep's own
-/// `register_package_schemas` call handles them when its manifest loads.
-/// No-op when the manifest declares no `schemas:` map.
-fn register_package_schemas(
-    project_path: &std::path::Path,
-    config: &crate::core::config::ProjectConfig,
-) -> Result<()> {
-    use crate::core::config::ProjectConfig;
-    use crate::core::embedded_schemas;
-    use streamlib_idents::SchemaEntry;
-
-    let Some(schemas) = config.schemas.as_ref() else {
-        return Ok(());
-    };
-
-    if schemas.is_empty() {
-        return Ok(());
-    }
-
-    let pkg_meta = config.package.as_ref().ok_or_else(|| {
-        Error::Configuration(format!(
-            "{} at {} declares `schemas:` but is missing a `package:` block. \
-             Schema canonical identifiers are composed from \
-             `package.{{org, name}}`.",
-            ProjectConfig::FILE_NAME,
-            project_path.display(),
-        ))
-    })?;
-
-    for (_name, entry) in schemas {
-        let SchemaEntry::Local { file } = entry else {
-            continue;
-        };
-        let schema_path = if file.is_absolute() {
-            file.clone()
-        } else {
-            project_path.join(file)
-        };
-        let body = std::fs::read_to_string(&schema_path).map_err(|e| {
-            Error::Configuration(format!(
-                "failed to read schema declared in {}: {}: {}",
-                ProjectConfig::FILE_NAME,
-                schema_path.display(),
-                e
-            ))
-        })?;
-        let canonical = canonical_identifier_for_schema(&body, pkg_meta, &schema_path)?;
-        tracing::debug!(
-            "registering schema '{}' from {}",
-            canonical,
-            schema_path.display()
-        );
-        embedded_schemas::register_schema(canonical, body);
-    }
-    Ok(())
-}
-
-/// Resolve a processor's bare-name config schema reference (#767) to
-/// its canonical id string. Walks the manifest's `schemas:` map via
-/// the supplied resolver output, locates the owning package + schema
-/// file, then reads the schema's `metadata.type` (new shape) or
-/// `metadata.name` (legacy reverse-DNS) to compose the canonical id.
-///
-/// `resolved` is `None` only when the project's processors all lack a
-/// `config:` block — reaching this function with `None` is a runtime
-/// bug.
-fn resolve_config_schema_canonical_id(
-    project_path: &std::path::Path,
-    config: &streamlib_processor_schema::ProcessorConfigSchema,
-    resolved: &Option<streamlib_idents::ResolvedPackages>,
-) -> std::result::Result<String, String> {
-    let resolved = resolved.as_ref().ok_or_else(|| {
-        format!(
-            "internal error: config-schema resolution requested but \
-             dependency graph for {} was not pre-resolved",
-            project_path.display()
-        )
-    })?;
-
-    let (owner, schema_path) = streamlib_idents::resolve_bare_schema_name(
-        resolved,
-        &resolved.root,
-        &config.schema,
-    )
-    .map_err(|e| format!("bare-name resolution failed: {}", e))?;
-
-    let owner_pkg = owner
-        .manifest
-        .package
-        .as_ref()
-        .ok_or_else(|| "owning package has no `package:` block".to_string())?;
-
-    // Read the schema's metadata to determine the canonical id form.
-    let body = std::fs::read_to_string(&schema_path)
-        .map_err(|e| format!("failed to read schema {}: {}", schema_path.display(), e))?;
-    let value: serde_yaml::Value = serde_yaml::from_str(&body)
-        .map_err(|e| format!("failed to parse schema {}: {}", schema_path.display(), e))?;
-    let metadata = value
-        .get("metadata")
-        .ok_or_else(|| format!("schema {} missing `metadata` block", schema_path.display()))?;
-
-    if let Some(type_str) = metadata.get("type").and_then(|t| t.as_str()) {
-        Ok(format!(
-            "@{}/{}/{}@{}",
-            owner_pkg.org.as_str(),
-            owner_pkg.name.as_str(),
-            type_str,
-            owner_pkg.version,
-        ))
-    } else if let Some(name_str) = metadata.get("name").and_then(|n| n.as_str()) {
-        // Legacy reverse-DNS form — append the owning package's semver.
-        Ok(format!("{}@{}", name_str, owner_pkg.version))
-    } else {
-        Err(format!(
-            "schema {} declares neither `metadata.type` nor `metadata.name`",
-            schema_path.display()
-        ))
-    }
-}
-
-/// Resolve a schema's canonical (unversioned) lookup key from its YAML
-/// body + the enclosing package's metadata. Mirrors
-/// `build.rs::resolve_canonical_identifier` so the engine const path
-/// and the runtime registration path produce identical keys.
-fn canonical_identifier_for_schema(
-    body: &str,
-    pkg_meta: &crate::core::config::PackageMetadata,
-    schema_path: &std::path::Path,
-) -> Result<String> {
-    let value: serde_yaml::Value = serde_yaml::from_str(body).map_err(|e| {
-        Error::Configuration(format!(
-            "failed to parse schema {}: {}",
-            schema_path.display(),
-            e
-        ))
-    })?;
-
-    let metadata = value.get("metadata").ok_or_else(|| {
-        Error::Configuration(format!(
-            "schema {} missing `metadata` block",
-            schema_path.display()
-        ))
-    })?;
-
-    if let Some(type_name) = metadata.get("type").and_then(|t| t.as_str()) {
-        return Ok(format!(
-            "@{}/{}/{}",
-            pkg_meta.org.as_str(),
-            pkg_meta.name.as_str(),
-            type_name
-        ));
-    }
-
-    if let Some(name) = metadata.get("name").and_then(|n| n.as_str()) {
-        return Ok(strip_legacy_semver_suffix(name).to_string());
-    }
-
-    Err(Error::Configuration(format!(
-        "schema {} must declare either `metadata.type` (new shape) or \
-         `metadata.name` (legacy reverse-DNS) — required for runtime \
-         registration",
-        schema_path.display()
-    )))
-}
-
-/// Strip a trailing `@MAJOR.MINOR.PATCH` semver suffix. Mirrors
-/// `embedded_schemas::strip_semver_suffix` (kept private here to avoid
-/// introducing a cross-module dep on a tiny string helper).
-fn strip_legacy_semver_suffix(name: &str) -> &str {
-    if let Some(at_pos) = name.rfind('@') {
-        let suffix = &name[at_pos + 1..];
-        let parts: Vec<&str> = suffix.split('.').collect();
-        if parts.len() == 3
-            && parts
-                .iter()
-                .all(|p| !p.is_empty() && p.chars().all(|c| c.is_ascii_digit()))
-        {
-            return &name[..at_pos];
-        }
-    }
-    name
-}
-
-/// The rustc target triple this engine was compiled for, captured at
-/// build time via `build.rs`. Same string Cargo prints for `--target`
-/// (e.g. `x86_64-unknown-linux-gnu`, `aarch64-apple-darwin`). Used to
-/// resolve plugin cdylibs inside `lib/<triple>/...` during package
-/// load — and exposed publicly so consumers (examples, application
-/// binaries) can write into the same triple-keyed directory layout
-/// without each running their own `build.rs`.
-pub fn host_target_triple() -> &'static str {
-    env!("STREAMLIB_HOST_TARGET")
-}
-
-/// Enumerate the target triples present as subdirectories of a
-/// package's `lib/` dir. Used in the error path when the host's triple
-/// has no matching artifact, so the user sees which platforms a slpkg
-/// WAS packed for instead of a generic "file not found".
-fn list_available_triples(lib_dir: &std::path::Path) -> Result<Vec<String>> {
-    if !lib_dir.is_dir() {
-        return Ok(Vec::new());
-    }
-    let entries = std::fs::read_dir(lib_dir).map_err(|e| {
-        Error::Configuration(format!(
-            "Failed to enumerate {}: {}",
-            lib_dir.display(),
-            e
-        ))
-    })?;
-    let mut triples: Vec<String> = entries
-        .filter_map(|entry| entry.ok())
-        .filter(|entry| entry.path().is_dir())
-        .filter_map(|entry| entry.file_name().to_str().map(str::to_string))
-        .collect();
-    triples.sort();
-    Ok(triples)
-}
-
-/// Extract a .slpkg ZIP archive to the package cache.
-/// Cache key is {name}-{version} from the embedded streamlib.yaml.
-/// Always overwrites on load.
-pub fn extract_slpkg_to_cache(slpkg_path: &std::path::Path) -> Result<std::path::PathBuf> {
-    use crate::core::config::ProjectConfig;
-
-    let slpkg_bytes = std::fs::read(slpkg_path).map_err(|e| {
-        Error::Configuration(format!("Failed to read {}: {}", slpkg_path.display(), e))
-    })?;
-
-    let cursor = std::io::Cursor::new(&slpkg_bytes);
-    let mut archive = zip::ZipArchive::new(cursor)
-        .map_err(|e| Error::Configuration(format!("Failed to open .slpkg archive: {}", e)))?;
-
-    // Read streamlib.yaml from archive to get name + version
-    let manifest_yaml = {
-        let mut manifest_file = archive.by_name(ProjectConfig::FILE_NAME).map_err(|e| {
-            Error::Configuration(format!(
-                ".slpkg archive missing {}: {}",
-                ProjectConfig::FILE_NAME,
-                e
-            ))
-        })?;
-        let mut contents = String::new();
-        std::io::Read::read_to_string(&mut manifest_file, &mut contents)
-            .map_err(|e| Error::Configuration(format!("Failed to read manifest: {}", e)))?;
-        contents
-    };
-
-    let config: ProjectConfig = serde_yaml::from_str(&manifest_yaml)
-        .map_err(|e| Error::Configuration(format!("Failed to parse manifest: {}", e)))?;
-
-    let package = config.package.as_ref().ok_or_else(|| {
-        Error::Configuration("streamlib.yaml missing [package] section".to_string())
-    })?;
-
-    let cache_key = format!("{}-{}", package.name, package.version);
-    let cache_dir = crate::core::streamlib_home::get_cached_package_dir(&cache_key);
-
-    // Always overwrite
-    if cache_dir.exists() {
-        std::fs::remove_dir_all(&cache_dir)
-            .map_err(|e| Error::Configuration(format!("Failed to clear cache dir: {}", e)))?;
-    }
-
-    tracing::info!(
-        "Extracting {} to {}",
-        slpkg_path.display(),
-        cache_dir.display()
-    );
-    std::fs::create_dir_all(&cache_dir)
-        .map_err(|e| Error::Configuration(format!("Failed to create cache dir: {}", e)))?;
-
-    // Re-open archive (cursor consumed by manifest read)
-    let cursor = std::io::Cursor::new(&slpkg_bytes);
-    let mut archive = zip::ZipArchive::new(cursor).map_err(|e| {
-        Error::Configuration(format!("Failed to re-open .slpkg archive: {}", e))
-    })?;
-
-    for i in 0..archive.len() {
-        let mut file = archive.by_index(i).map_err(|e| {
-            Error::Configuration(format!("Failed to read archive entry: {}", e))
-        })?;
-
-        let file_name = file.name().to_string();
-
-        // Security: reject path traversal
-        if file_name.contains("..") || file_name.starts_with('/') {
-            return Err(Error::Configuration(format!(
-                "Invalid path in .slpkg archive: {}",
-                file_name
-            )));
-        }
-
-        let output_path = cache_dir.join(&file_name);
-
-        if let Some(parent) = output_path.parent() {
-            std::fs::create_dir_all(parent).map_err(|e| {
-                Error::Configuration(format!("Failed to create directory: {}", e))
-            })?;
-        }
-
-        let mut output_file = std::fs::File::create(&output_path).map_err(|e| {
-            Error::Configuration(format!("Failed to create {}: {}", output_path.display(), e))
-        })?;
-
-        std::io::copy(&mut file, &mut output_file).map_err(|e| {
-            Error::Configuration(format!("Failed to extract {}: {}", file_name, e))
-        })?;
-    }
-
-    Ok(cache_dir)
-}
 
 /// Compute the per-runtime surface-sharing socket path, refuse to start if
 /// another live runtime is already bound there, clean up an orphan socket
@@ -2249,697 +1072,6 @@ fn bring_up_surface_service(
     Ok((Arc::new(Mutex::new(Some(service))), socket_path))
 }
 
-// =============================================================================
-// ModuleResolverStrategy — unified resolver for the four module-loading
-// entry points. Each variant covers one historical loader behavior;
-// `add_module_with(ident, strategy)` is the single dispatch point.
-// =============================================================================
-
-/// How [`Runner::add_module_with`] should source the manifest for a
-/// module. Each variant maps to one of the historical loader methods
-/// (`load_project`, `load_package`, `load_workspace_packages`) plus the
-/// imperative default chain that bare [`Runner::add_module`] uses.
-///
-/// The strategy is the in-code equivalent of the manifest's `patch:`
-/// table — callers can pin a dep to a workspace stage dir, a specific
-/// resolver tier, a `.slpkg` archive, or an arbitrary directory
-/// without editing yaml.
-#[derive(Debug, Clone)]
-pub enum ModuleResolverStrategy {
-    /// Default ident-keyed resolution chain: workspace stage dir
-    /// (`<workspace>/target/streamlib-plugins/<org>__<name>/`) first,
-    /// falling back to the installed-package cache. Behavior used by
-    /// bare [`Runner::add_module`] and by transitive `dependencies:`
-    /// entries declared as `"@org/name": "^version"`.
-    DefaultChain,
-
-    /// Look up the workspace stage dir only; do not fall back to the
-    /// installed-package cache. Counterpart of the legacy
-    /// [`Runner::load_workspace_packages`] per-id lookup.
-    WorkspaceStaged,
-
-    /// Look up the installed-package cache only; skip workspace.
-    /// Useful for runtimes that explicitly want the installed-from-
-    /// slpkg copy even when a workspace stage dir exists.
-    InstalledCache,
-
-    /// Load the manifest at this directory directly. Counterpart of
-    /// the legacy [`Runner::load_project`].
-    ManifestDirectory { path: std::path::PathBuf },
-
-    /// Extract this `.slpkg` archive into the package cache, then
-    /// load the extracted manifest. Counterpart of the legacy
-    /// [`Runner::load_package`].
-    SlpkgArchive { path: std::path::PathBuf },
-}
-
-/// Resolve a [`ModuleResolverStrategy`] to `(manifest_dir, on_disk_version)`.
-///
-/// `package_ref` supplies the canonical `@org/name` for the
-/// ident-keyed strategies ([`ModuleResolverStrategy::DefaultChain`],
-/// [`ModuleResolverStrategy::WorkspaceStaged`],
-/// [`ModuleResolverStrategy::InstalledCache`]); it is ignored by the
-/// path-keyed ones.
-///
-/// The returned version is sourced from the resolver tier that hit:
-/// workspace stage dirs re-parse the staged manifest, the installed
-/// cache reads `InstalledPackageEntry::version`, the path-keyed
-/// variants read the manifest at the directory. The caller is
-/// responsible for validating it against the requested
-/// [`SemVerRange`] (see [`Runner::add_module_recursively`]).
-fn resolve_strategy_to_manifest_dir(
-    strategy: &ModuleResolverStrategy,
-    package_ref: Option<&streamlib_idents::PackageRef>,
-) -> std::result::Result<(std::path::PathBuf, streamlib_idents::SemVer), AddModuleError> {
-    match strategy {
-        ModuleResolverStrategy::DefaultChain => {
-            let pkg_ref = package_ref.ok_or_else(|| AddModuleError::StrategyNeedsPackageRef {
-                strategy: "DefaultChain".into(),
-            })?;
-            // Try workspace stage dir first, then installed cache.
-            if let Some(workspace_root) = locate_workspace_root_for_strategy()? {
-                let staged_dir = workspace_stage_dir(&workspace_root, pkg_ref);
-                if staged_dir.join("streamlib.yaml").exists() {
-                    return Ok((staged_dir.clone(), read_version_from_manifest_dir(&staged_dir)?));
-                }
-            }
-            lookup_installed_cache(pkg_ref)?
-                .ok_or_else(|| AddModuleError::ModuleNotFound { package: pkg_ref.clone() })
-        }
-        ModuleResolverStrategy::WorkspaceStaged => {
-            let pkg_ref = package_ref.ok_or_else(|| AddModuleError::StrategyNeedsPackageRef {
-                strategy: "WorkspaceStaged".into(),
-            })?;
-            let workspace_root = locate_workspace_root_for_strategy()?
-                .ok_or(AddModuleError::WorkspaceRootNotFound)?;
-            let staged_dir = workspace_stage_dir(&workspace_root, pkg_ref);
-            if !staged_dir.join("streamlib.yaml").exists() {
-                return Err(AddModuleError::WorkspaceStageMiss {
-                    package: pkg_ref.clone(),
-                    expected_path: staged_dir,
-                });
-            }
-            // For Rust-impl packages, the cdylib must exist at
-            // `lib/<host_triple>/`. Surface the missing-cdylib case
-            // explicitly so callers see the actionable diagnostic
-            // (rather than letting `register_manifest_processors` fail
-            // with a generic "missing dylib for this triple" message).
-            check_cdylib_present_when_rust_impl(&staged_dir, pkg_ref)?;
-            let version = read_version_from_manifest_dir(&staged_dir)?;
-            Ok((staged_dir, version))
-        }
-        ModuleResolverStrategy::InstalledCache => {
-            let pkg_ref = package_ref.ok_or_else(|| AddModuleError::StrategyNeedsPackageRef {
-                strategy: "InstalledCache".into(),
-            })?;
-            lookup_installed_cache(pkg_ref)?
-                .ok_or_else(|| AddModuleError::ModuleNotFound { package: pkg_ref.clone() })
-        }
-        ModuleResolverStrategy::ManifestDirectory { path } => {
-            if !path.join("streamlib.yaml").exists() {
-                return Err(AddModuleError::ManifestDirectoryMissing { path: path.clone() });
-            }
-            let version = read_version_from_manifest_dir(path)?;
-            Ok((path.clone(), version))
-        }
-        ModuleResolverStrategy::SlpkgArchive { path } => {
-            let extracted =
-                extract_slpkg_to_cache(path).map_err(|e| AddModuleError::SlpkgExtractionFailed {
-                    archive: path.clone(),
-                    detail: e.to_string(),
-                })?;
-            let version = read_version_from_manifest_dir(&extracted)?;
-            Ok((extracted, version))
-        }
-    }
-}
-
-/// Look the canonical [`streamlib_idents::PackageRef`] up in the
-/// installed-package cache (`InstalledPackageManifest`). Returns the
-/// `(cache_dir, version)` pair when present.
-fn lookup_installed_cache(
-    pkg_ref: &streamlib_idents::PackageRef,
-) -> std::result::Result<Option<(std::path::PathBuf, streamlib_idents::SemVer)>, AddModuleError> {
-    use crate::core::config::InstalledPackageManifest;
-    use crate::core::streamlib_home::get_cached_package_dir;
-
-    let manifest =
-        InstalledPackageManifest::load().map_err(|e| AddModuleError::InstalledCacheLoadFailed {
-            detail: e.to_string(),
-        })?;
-    Ok(manifest
-        .find_by_ref(pkg_ref)
-        .map(|entry| (get_cached_package_dir(&entry.cache_dir), entry.version)))
-}
-
-/// Best-effort workspace-root resolution for strategy lookups.
-/// Honors `STREAMLIB_WORKSPACE_ROOT` strictly (typo'd env var ⇒
-/// `WorkspaceRootInvalid`); otherwise falls back to
-/// `cargo locate-project --workspace`, returning `Ok(None)` when no
-/// workspace is reachable (so [`ModuleResolverStrategy::DefaultChain`]
-/// can silently fall through to the installed cache).
-fn locate_workspace_root_for_strategy(
-) -> std::result::Result<Option<std::path::PathBuf>, AddModuleError> {
-    if let Ok(env_root) = std::env::var("STREAMLIB_WORKSPACE_ROOT") {
-        let path = std::path::PathBuf::from(&env_root);
-        if !path.is_dir() {
-            return Err(AddModuleError::WorkspaceRootInvalid { env_value: env_root });
-        }
-        return Ok(Some(path));
-    }
-    Ok(resolve_workspace_root().ok())
-}
-
-fn workspace_stage_dir(
-    workspace_root: &std::path::Path,
-    pkg_ref: &streamlib_idents::PackageRef,
-) -> std::path::PathBuf {
-    workspace_root
-        .join("target")
-        .join("streamlib-plugins")
-        .join(format!(
-            "{}__{}",
-            pkg_ref.org.as_str(),
-            pkg_ref.name.as_str()
-        ))
-}
-
-/// Read the `[package].version` field from the streamlib.yaml at
-/// `dir`. Surfaces [`AddModuleError::ManifestDirectoryMissing`] when
-/// the file is absent and a generic [`AddModuleError::ManifestLoadFailed`]
-/// (with an `any` ident placeholder; the caller re-wraps with the
-/// real ident) when the parse or package-block lookup fails.
-fn read_version_from_manifest_dir(
-    dir: &std::path::Path,
-) -> std::result::Result<streamlib_idents::SemVer, AddModuleError> {
-    use streamlib_idents::Manifest;
-    let manifest_path = dir.join(Manifest::FILE_NAME);
-    if !manifest_path.exists() {
-        return Err(AddModuleError::ManifestDirectoryMissing { path: dir.to_path_buf() });
-    }
-    let manifest = Manifest::load(dir).map_err(|e| AddModuleError::StrategyManifestLoadFailed {
-        source_path: dir.to_path_buf(),
-        detail: e.to_string(),
-    })?;
-    manifest
-        .package
-        .as_ref()
-        .map(|p| p.version)
-        .ok_or_else(|| AddModuleError::StrategyManifestLoadFailed {
-            source_path: dir.to_path_buf(),
-            detail: "manifest has no `package:` block".into(),
-        })
-}
-
-/// Read just the `(org, name, version)` triple from a manifest
-/// directory's streamlib.yaml and build an `any`-version
-/// [`ModuleIdent`] from it. Used by the path-keyed wrappers
-/// ([`Runner::load_project`], [`Runner::load_package`]) to build the
-/// strict ident the unified flow validates against.
-fn read_module_ident_from_manifest_dir(
-    dir: &std::path::Path,
-) -> Result<streamlib_idents::ModuleIdent> {
-    use crate::core::config::ProjectConfig;
-    let config = ProjectConfig::load(dir)?;
-    let pkg = config.package.as_ref().ok_or_else(|| {
-        Error::Configuration(format!(
-            "{} at {} has no `[package]` block — required to build a ModuleIdent.",
-            ProjectConfig::FILE_NAME,
-            dir.display()
-        ))
-    })?;
-    Ok(streamlib_idents::ModuleIdent::any(pkg.org.clone(), pkg.name.clone()))
-}
-
-/// Peek the `[package]` block out of a `.slpkg` archive's embedded
-/// manifest WITHOUT fully extracting the archive. Used by
-/// [`Runner::load_package`] to build the ident the unified flow
-/// validates against — the strategy's extraction step then writes
-/// the package to the cache.
-fn read_module_ident_from_slpkg(
-    slpkg_path: &std::path::Path,
-) -> Result<streamlib_idents::ModuleIdent> {
-    use crate::core::config::ProjectConfig;
-    let bytes = std::fs::read(slpkg_path)
-        .map_err(|e| Error::Configuration(format!("Failed to read {}: {}", slpkg_path.display(), e)))?;
-    let mut archive = zip::ZipArchive::new(std::io::Cursor::new(&bytes))
-        .map_err(|e| Error::Configuration(format!("Failed to open .slpkg archive: {}", e)))?;
-    let mut manifest_file = archive
-        .by_name(ProjectConfig::FILE_NAME)
-        .map_err(|e| Error::Configuration(format!(".slpkg archive missing {}: {}", ProjectConfig::FILE_NAME, e)))?;
-    let mut yaml_body = String::new();
-    std::io::Read::read_to_string(&mut manifest_file, &mut yaml_body)
-        .map_err(|e| Error::Configuration(format!("Failed to read manifest from .slpkg: {}", e)))?;
-    let config: ProjectConfig = serde_yaml::from_str(&yaml_body)
-        .map_err(|e| Error::Configuration(format!("Failed to parse manifest from .slpkg: {}", e)))?;
-    let pkg = config.package.as_ref().ok_or_else(|| {
-        Error::Configuration(format!(
-            ".slpkg at {} has no `[package]` block — required to build a ModuleIdent.",
-            slpkg_path.display()
-        ))
-    })?;
-    Ok(streamlib_idents::ModuleIdent::any(pkg.org.clone(), pkg.name.clone()))
-}
-
-/// Check that a Rust-impl staged package has its cdylib present at
-/// `lib/<host_triple>/`. No-op for schemas-only or non-Rust packages.
-/// Used by [`ModuleResolverStrategy::WorkspaceStaged`] to surface the
-/// missing-cdylib case as [`AddModuleError::CdylibMissingForRustImpl`]
-/// rather than letting `register_manifest_processors` fail later with
-/// a less actionable message.
-fn check_cdylib_present_when_rust_impl(
-    staged_dir: &std::path::Path,
-    pkg_ref: &streamlib_idents::PackageRef,
-) -> std::result::Result<(), AddModuleError> {
-    let body = match std::fs::read_to_string(staged_dir.join("streamlib.yaml")) {
-        Ok(b) => b,
-        Err(_) => return Ok(()),
-    };
-    let manifest: streamlib_processor_schema::ProjectConfigMinimal =
-        match serde_yaml::from_str(&body) {
-            Ok(m) => m,
-            Err(_) => return Ok(()),
-        };
-    let has_rust = manifest.processors.iter().any(|p| {
-        matches!(
-            p.runtime.language,
-            streamlib_processor_schema::ProcessorLanguage::Rust
-        )
-    });
-    if !has_rust {
-        return Ok(());
-    }
-    let triple_dir = staged_dir.join("lib").join(host_target_triple());
-    let dylib_ext = if cfg!(target_os = "macos") {
-        "dylib"
-    } else if cfg!(target_os = "windows") {
-        "dll"
-    } else {
-        "so"
-    };
-    let any_dylib_present = std::fs::read_dir(&triple_dir)
-        .map(|iter| {
-            iter.flatten()
-                .any(|e| e.path().extension().is_some_and(|ext| ext == dylib_ext))
-        })
-        .unwrap_or(false);
-    if !any_dylib_present {
-        return Err(AddModuleError::CdylibMissingForRustImpl {
-            package: pkg_ref.clone(),
-            expected_path: triple_dir,
-        });
-    }
-    Ok(())
-}
-
-/// Bridge helper used by `load_workspace_packages`'s back-compat
-/// error translation: split an `"@org/name"` actual-id string back
-/// into `(org, name)` halves so the legacy
-/// `PackageIdentityMismatch` variant can carry the structured fields
-/// the historical tests assert on. Falls back to empty strings when
-/// the input shape is unexpected — the wrapper only uses this on
-/// inputs that the unified flow itself produced.
-fn split_canonical_for_legacy_error(actual: &str) -> (String, String) {
-    actual
-        .strip_prefix('@')
-        .and_then(|s| s.split_once('/'))
-        .map(|(o, n)| (o.to_string(), n.to_string()))
-        .unwrap_or_default()
-}
-
-/// Per-failure-mode error returned by [`Runner::add_module`].
-#[derive(Debug, thiserror::Error)]
-pub enum AddModuleError {
-    /// No workspace stage dir AND no installed-package cache entry
-    /// matches `@org/name`. Surface the canonical ref so callers can
-    /// suggest `streamlib pkg install`, `cargo xtask build-plugins`,
-    /// or a typo fix.
-    #[error(
-        "Module '{package}' not found — no workspace stage dir at \
-         `target/streamlib-plugins/<org>__<name>/` and no installed-cache \
-         entry. Run `cargo xtask build-plugins --package {package}` (dev) \
-         or `streamlib pkg install <slpkg>` (distribution)."
-    )]
-    ModuleNotFound {
-        package: streamlib_idents::PackageRef,
-    },
-
-    /// Workspace stage dir or installed-cache entry was found but the
-    /// `streamlib.yaml` failed to parse / lacked a `package:` block.
-    #[error(
-        "Failed to load manifest for '{module}' from {}: {detail}",
-        source_path.display()
-    )]
-    ManifestLoadFailed {
-        module: streamlib_idents::ModuleIdent,
-        source_path: std::path::PathBuf,
-        detail: String,
-    },
-
-    /// Workspace stage dir was found but its `streamlib.yaml`'s
-    /// `package: { org, name }` doesn't match the requested ident
-    /// (manual clobbering, stale rename, wrong dir).
-    #[error(
-        "Module '{module}' identity mismatch at {}: \
-         staged manifest declares `{actual}`. \
-         Re-run `cargo xtask build-plugins` to regenerate.",
-        source_path.display()
-    )]
-    ManifestIdentityMismatch {
-        module: streamlib_idents::ModuleIdent,
-        source_path: std::path::PathBuf,
-        actual: String,
-    },
-
-    /// On-disk version doesn't satisfy the ident's [`SemVerRange`].
-    #[error(
-        "Module '{module}' resolved to version {found} at {} which doesn't \
-         satisfy the requested range. Install a matching version or relax \
-         the range.",
-        source_path.display()
-    )]
-    VersionRangeUnsatisfied {
-        module: streamlib_idents::ModuleIdent,
-        found: streamlib_idents::SemVer,
-        source_path: std::path::PathBuf,
-    },
-
-    /// `InstalledPackageManifest::load()` errored before lookup could
-    /// run. Catches I/O / parse failures distinct from "no entry."
-    #[error("Failed to load installed-package cache: {detail}")]
-    InstalledCacheLoadFailed { detail: String },
-
-    /// `STREAMLIB_WORKSPACE_ROOT` was set but its value isn't a
-    /// directory. Treats the env var as the user's stated intent —
-    /// don't silently fall through to the installed cache when the
-    /// override is broken. Mirrors `load_workspace_packages`'s
-    /// `WorkspaceRootNotFound` behavior on env-var-set-but-invalid.
-    #[error(
-        "STREAMLIB_WORKSPACE_ROOT is set to `{env_value}` but that path \
-         is not a directory. Fix the env var or unset it (in which case \
-         the resolver falls through to the installed-package cache)."
-    )]
-    WorkspaceRootInvalid { env_value: String },
-
-    /// `Runner::load_project` (or the recursive dep walker rooted in
-    /// it) rejected the resolved source path. Wraps the underlying
-    /// engine [`Error`] so callers can introspect.
-    #[error("load_project failed for '{module}': {source}")]
-    LoadProjectFailed {
-        module: streamlib_idents::ModuleIdent,
-        #[source]
-        source: Box<Error>,
-    },
-
-    /// `Runner::add_module_with` was called with
-    /// [`ModuleResolverStrategy::WorkspaceStaged`] but no streamlib.yaml
-    /// exists at the workspace stage dir. Surface the expected path so
-    /// callers see exactly where the resolver looked.
-    #[error(
-        "Module '{package}' not staged under target/streamlib-plugins. \
-         Expected `streamlib.yaml` at {expected_path}. \
-         Run `cargo xtask build-plugins --package {package}` first."
-    )]
-    WorkspaceStageMiss {
-        package: streamlib_idents::PackageRef,
-        expected_path: std::path::PathBuf,
-    },
-
-    /// Workspace stage dir resolution requires a workspace root but
-    /// neither `STREAMLIB_WORKSPACE_ROOT` nor `cargo locate-project`
-    /// returned one. Distinct from
-    /// [`Self::WorkspaceRootInvalid`] (set-but-broken env var).
-    #[error(
-        "Workspace root not found — set STREAMLIB_WORKSPACE_ROOT or run \
-         from within a Cargo workspace"
-    )]
-    WorkspaceRootNotFound,
-
-    /// A `Rust`-impl workspace-staged package has no cdylib at
-    /// `lib/<host_triple>/`. The staged manifest declares Rust
-    /// processors but `cargo xtask build-plugins` either was never
-    /// run or produced no artifact for this host triple.
-    #[error(
-        "Cdylib missing for Rust-impl package '{package}' — expected at \
-         {expected_path}. Re-run `cargo xtask build-plugins` to rebuild."
-    )]
-    CdylibMissingForRustImpl {
-        package: streamlib_idents::PackageRef,
-        expected_path: std::path::PathBuf,
-    },
-
-    /// [`ModuleResolverStrategy::ManifestDirectory`] pointed at a
-    /// directory that does not contain a `streamlib.yaml`. Catches
-    /// the `load_project("./does-not-exist")` and patch-points-at-
-    /// missing-path cases at the strategy layer.
-    #[error("Manifest directory has no streamlib.yaml at {}", path.display())]
-    ManifestDirectoryMissing { path: std::path::PathBuf },
-
-    /// Strategy was [`ModuleResolverStrategy::SlpkgArchive`] and the
-    /// extraction step failed (I/O, malformed ZIP, missing embedded
-    /// manifest, etc.).
-    #[error(
-        "Failed to extract .slpkg archive at {}: {detail}",
-        archive.display()
-    )]
-    SlpkgExtractionFailed {
-        archive: std::path::PathBuf,
-        detail: String,
-    },
-
-    /// Strategy resolver failed while reading the manifest at the
-    /// resolved directory (parse error, missing `[package]` block).
-    /// Distinct from [`Self::ManifestLoadFailed`] because the caller
-    /// hasn't bound a [`ModuleIdent`] yet at this stage.
-    #[error(
-        "Failed to read manifest at {}: {detail}",
-        source_path.display()
-    )]
-    StrategyManifestLoadFailed {
-        source_path: std::path::PathBuf,
-        detail: String,
-    },
-
-    /// Strategy was ident-keyed
-    /// ([`ModuleResolverStrategy::DefaultChain`] /
-    /// [`ModuleResolverStrategy::WorkspaceStaged`] /
-    /// [`ModuleResolverStrategy::InstalledCache`]) but no
-    /// [`PackageRef`](streamlib_idents::PackageRef) was supplied.
-    /// Internal invariant — callers route through
-    /// [`Runner::add_module_with`] which always supplies the ref.
-    #[error(
-        "Strategy '{strategy}' requires a PackageRef but none was supplied — \
-         internal invariant violation"
-    )]
-    StrategyNeedsPackageRef { strategy: String },
-
-    /// A dependency cycle was detected during recursive dep walking.
-    /// `cycle` lists the [`PackageRef`](streamlib_idents::PackageRef)
-    /// that was hit while already mid-load on the recursion stack.
-    /// Cycles can only arise from a `streamlib.yaml` that transitively
-    /// depends on itself; surface the canonical ref so the dev can
-    /// trace the offending edge.
-    #[error(
-        "Dependency cycle detected — package {} is already mid-load on the \
-         recursion stack while a transitive dep tries to load it again",
-        cycle.iter().map(|p| p.to_string()).collect::<Vec<_>>().join(" → ")
-    )]
-    DependencyCycleDetected {
-        cycle: Vec<streamlib_idents::PackageRef>,
-    },
-}
-
-impl From<AddModuleError> for Error {
-    fn from(err: AddModuleError) -> Self {
-        match err {
-            AddModuleError::LoadProjectFailed { source, .. } => *source,
-            other => Error::Configuration(other.to_string()),
-        }
-    }
-}
-
-/// Per-failure-mode error returned by [`Runner::remove_module`].
-///
-/// Today the only variant is the milestone-deferral marker; new
-/// variants land when the hot-reload lifecycle work ships.
-#[derive(Debug, thiserror::Error)]
-pub enum RemoveModuleError {
-    /// Module unload requires the hot-reload lifecycle work that's
-    /// explicitly out of scope for the current All-Dynamic Package
-    /// Loading milestone. Calling `remove_module` returns this without
-    /// altering any runtime state — see [`Runner::remove_module`] for
-    /// the rationale.
-    #[error(
-        "remove_module('{module}') is not yet implemented — \
-         hot-reload lifecycle is deferred to a future milestone. \
-         The runtime currently supports load-only, runtime-lifetime \
-         module registration."
-    )]
-    HotReloadLifecycleNotYetImplemented {
-        module: streamlib_idents::ModuleIdent,
-    },
-}
-
-impl From<RemoveModuleError> for Error {
-    fn from(err: RemoveModuleError) -> Self {
-        Error::Configuration(err.to_string())
-    }
-}
-
-// =============================================================================
-// load_workspace_packages — typed error + helpers
-// =============================================================================
-
-/// Per-failure-mode error returned by [`Runner::load_workspace_packages`].
-///
-/// The variants surface enough context (offending name, expected path,
-/// underlying engine error) that callers can match for retry vs. abort
-/// or surface an actionable message to the developer.
-#[derive(Debug, thiserror::Error)]
-pub enum LoadWorkspacePackagesError {
-    /// Name did not parse as `@<org>/<name>` per the typed `streamlib-idents`
-    /// org / name validators (charset, leading-letter, length).
-    #[error("Invalid package id '{0}' — expected `@<org>/<name>` with lowercase org and name")]
-    InvalidPackageId(String),
-
-    /// Workspace root could not be resolved — neither the
-    /// `STREAMLIB_WORKSPACE_ROOT` env var nor `cargo locate-project`
-    /// returned a usable directory.
-    #[error(
-        "Workspace root not found — set STREAMLIB_WORKSPACE_ROOT or run \
-         from within a Cargo workspace"
-    )]
-    WorkspaceRootNotFound,
-
-    /// Staged dir does not exist for this package. Most likely cause:
-    /// the dev hasn't run `cargo xtask build-plugins` yet (or pruned
-    /// `target/` since the last run).
-    #[error(
-        "Package '{name}' not staged at {expected_path}. \
-         Run `cargo xtask build-plugins` first."
-    )]
-    PackageNotStaged {
-        name: String,
-        expected_path: std::path::PathBuf,
-    },
-
-    /// Staged dir exists and parses, but its `[package]` org / name
-    /// don't match the requested id. Catches the case where the
-    /// staged tree was clobbered out-of-band (manual `cp`, stale
-    /// rename) before the runtime registers the wrong processors.
-    #[error(
-        "Package identity mismatch at {staged_path}: \
-         requested `@{requested_org}/{requested_name}`, found \
-         `@{actual_org}/{actual_name}` in staged streamlib.yaml. \
-         Re-run `cargo xtask build-plugins` to regenerate."
-    )]
-    PackageIdentityMismatch {
-        staged_path: std::path::PathBuf,
-        requested_org: String,
-        requested_name: String,
-        actual_org: String,
-        actual_name: String,
-    },
-
-    /// Staged dir is present and identity matches, but a Rust-impl
-    /// package's expected cdylib is missing under `lib/<host_triple>/`.
-    /// Distinguishes "staging succeeded but cargo build silently
-    /// produced no artifact" from a generic load_project failure.
-    #[error(
-        "Cdylib missing for Rust-impl package '{name}' — expected at \
-         {expected_path}. Re-run `cargo xtask build-plugins` to rebuild."
-    )]
-    CdylibMissing {
-        name: String,
-        expected_path: std::path::PathBuf,
-    },
-
-    /// `load_project` rejected the staged dir. Carries the engine
-    /// `Error` so callers can introspect further.
-    #[error("load_project failed for '{name}': {source}")]
-    LoadProjectFailed {
-        name: String,
-        #[source]
-        source: Box<Error>,
-    },
-}
-
-impl From<LoadWorkspacePackagesError> for Error {
-    fn from(err: LoadWorkspacePackagesError) -> Self {
-        match err {
-            LoadWorkspacePackagesError::LoadProjectFailed { source, .. } => *source,
-            other => Error::Configuration(other.to_string()),
-        }
-    }
-}
-
-/// Parsed `@<org>/<name>` canonical id. Only the slice references are
-/// kept — the caller's input must outlive this struct.
-#[derive(Debug)]
-struct CanonicalPackageId<'a> {
-    org_str: &'a str,
-    name_str: &'a str,
-}
-
-fn parse_canonical_package_id(name: &str) -> std::result::Result<
-    CanonicalPackageId<'_>,
-    LoadWorkspacePackagesError,
-> {
-    // Strip the leading '@', split on first '/', then route the
-    // halves through the typed `streamlib-idents` validators so
-    // charset / leading-letter / length rules apply here too. We
-    // don't surface the typed-parser's stringy diagnostic — the
-    // typed `InvalidPackageId` variant is what callers match on —
-    // but using the validators means a name like `@TaToLaB/CAMERA`
-    // fails fast here rather than at the filesystem-lookup stage.
-    let stripped = name
-        .strip_prefix('@')
-        .ok_or_else(|| LoadWorkspacePackagesError::InvalidPackageId(name.to_string()))?;
-    let (org, pkg) = stripped
-        .split_once('/')
-        .ok_or_else(|| LoadWorkspacePackagesError::InvalidPackageId(name.to_string()))?;
-    if org.is_empty() || pkg.is_empty() || pkg.contains('/') {
-        return Err(LoadWorkspacePackagesError::InvalidPackageId(name.to_string()));
-    }
-    streamlib_idents::Org::new(org)
-        .map_err(|_| LoadWorkspacePackagesError::InvalidPackageId(name.to_string()))?;
-    streamlib_idents::Package::new(pkg)
-        .map_err(|_| LoadWorkspacePackagesError::InvalidPackageId(name.to_string()))?;
-    Ok(CanonicalPackageId { org_str: org, name_str: pkg })
-}
-
-fn resolve_workspace_root() -> std::result::Result<std::path::PathBuf, LoadWorkspacePackagesError> {
-    // Env-var override wins when set AND the path resolves — the env
-    // var IS the user's intent, so a typo'd path should surface as a
-    // precise error rather than silently falling through to cargo.
-    if let Ok(env_root) = std::env::var("STREAMLIB_WORKSPACE_ROOT") {
-        let path = std::path::PathBuf::from(&env_root);
-        return if path.is_dir() {
-            Ok(path)
-        } else {
-            Err(LoadWorkspacePackagesError::WorkspaceRootNotFound)
-        };
-    }
-
-    let output = std::process::Command::new("cargo")
-        .args(["locate-project", "--workspace", "--message-format=plain"])
-        .output()
-        .map_err(|_| LoadWorkspacePackagesError::WorkspaceRootNotFound)?;
-    if !output.status.success() {
-        return Err(LoadWorkspacePackagesError::WorkspaceRootNotFound);
-    }
-    let manifest_path = String::from_utf8(output.stdout)
-        .map_err(|_| LoadWorkspacePackagesError::WorkspaceRootNotFound)?;
-    let trimmed = manifest_path.trim();
-    if trimmed.is_empty() {
-        return Err(LoadWorkspacePackagesError::WorkspaceRootNotFound);
-    }
-    std::path::PathBuf::from(trimmed)
-        .parent()
-        .map(|p| p.to_path_buf())
-        .ok_or(LoadWorkspacePackagesError::WorkspaceRootNotFound)
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -2991,190 +1123,6 @@ mod tests {
         );
     }
 
-    #[test]
-    fn parse_canonical_package_id_accepts_well_formed_input() {
-        // Tightest happy-path lock: every component round-trips
-        // (post-`@`, pre-`/`, post-`/`) into the parsed slices. The
-        // parser is the contract for what `load_workspace_packages`
-        // treats as a legal id — a regression that flipped `org` and
-        // `name` would break the lookup silently.
-        let parsed = parse_canonical_package_id("@tatolab/camera").unwrap();
-        assert_eq!(parsed.org_str, "tatolab");
-        assert_eq!(parsed.name_str, "camera");
-    }
-
-    #[test]
-    fn parse_canonical_package_id_rejects_missing_at_prefix() {
-        let err = parse_canonical_package_id("tatolab/camera").unwrap_err();
-        assert!(matches!(
-            err,
-            LoadWorkspacePackagesError::InvalidPackageId(ref s) if s == "tatolab/camera"
-        ));
-    }
-
-    #[test]
-    fn parse_canonical_package_id_rejects_missing_slash() {
-        let err = parse_canonical_package_id("@tatolab").unwrap_err();
-        assert!(matches!(
-            err,
-            LoadWorkspacePackagesError::InvalidPackageId(ref s) if s == "@tatolab"
-        ));
-    }
-
-    #[test]
-    fn parse_canonical_package_id_rejects_empty_org_or_name() {
-        // Both halves of `@<org>/<name>` must be non-empty. Mentally
-        // reverting either non-empty check would let `@/foo` or
-        // `@foo/` through to the path resolver, where the lookup
-        // would produce a confusing `not_staged` error instead of
-        // the precise `InvalidPackageId` one.
-        for bad in ["@/camera", "@tatolab/", "@/"] {
-            let err = parse_canonical_package_id(bad).unwrap_err();
-            assert!(matches!(err, LoadWorkspacePackagesError::InvalidPackageId(_)));
-        }
-    }
-
-    #[test]
-    fn parse_canonical_package_id_rejects_extra_slashes() {
-        // `@org/name` is a 2-segment id; nesting (`@org/group/name`)
-        // is not a legal shape. The lookup format must match the
-        // staged dir name `<org>__<name>` which is two-segment by
-        // construction.
-        let err = parse_canonical_package_id("@org/sub/name").unwrap_err();
-        assert!(matches!(err, LoadWorkspacePackagesError::InvalidPackageId(_)));
-    }
-
-    #[test]
-    #[serial]
-    fn resolve_workspace_root_honors_streamlib_workspace_root_env_var() {
-        // Test fixture: tempdir set via STREAMLIB_WORKSPACE_ROOT must
-        // win over the cargo-locate-project fallback. Mentally
-        // reverting the env-var branch would silently fall through to
-        // cargo and pick the streamlib workspace root, which is
-        // wrong for hosting CI fixtures.
-        let tmp = tempfile::tempdir().unwrap();
-        let key = "STREAMLIB_WORKSPACE_ROOT";
-        let prev = std::env::var_os(key);
-        // SAFETY: protected by `#[serial]` against parallel test mutation.
-        unsafe {
-            std::env::set_var(key, tmp.path());
-        }
-        let resolved = resolve_workspace_root().unwrap();
-        unsafe {
-            match prev {
-                Some(v) => std::env::set_var(key, v),
-                None => std::env::remove_var(key),
-            }
-        }
-        assert_eq!(resolved, tmp.path());
-    }
-
-    #[test]
-    #[serial]
-    fn resolve_workspace_root_errors_when_env_var_path_does_not_exist() {
-        // Env var IS the user's stated intent — when it points at a
-        // non-existent dir, surface `WorkspaceRootNotFound` directly
-        // rather than silently falling through to cargo locate-project.
-        // The previous "fall through" branch swallowed user typos as
-        // "ran from outside a workspace" errors; this test pins the
-        // new behavior so a regression to silent-fallthrough fails.
-        let key = "STREAMLIB_WORKSPACE_ROOT";
-        let prev = std::env::var_os(key);
-        unsafe {
-            std::env::set_var(key, "/nonexistent/path/that/does/not/exist");
-        }
-        let err = resolve_workspace_root().unwrap_err();
-        unsafe {
-            match prev {
-                Some(v) => std::env::set_var(key, v),
-                None => std::env::remove_var(key),
-            }
-        }
-        assert!(matches!(err, LoadWorkspacePackagesError::WorkspaceRootNotFound));
-    }
-
-    #[test]
-    fn parse_canonical_package_id_rejects_uppercase_via_typed_validator() {
-        // The typed `streamlib-idents` validators enforce lowercase
-        // for org/name. Bypass would let `@TaToLaB/CAMERA` parse
-        // here, hit the filesystem at `target/streamlib-plugins/TaToLaB__CAMERA`,
-        // and surface a less informative `PackageNotStaged` error.
-        // Reverting the `Org::new` / `Package::new` validators would
-        // pass this through; the test catches that regression.
-        let err = parse_canonical_package_id("@TaToLaB/camera").unwrap_err();
-        assert!(matches!(err, LoadWorkspacePackagesError::InvalidPackageId(_)));
-        let err = parse_canonical_package_id("@tatolab/CAMERA").unwrap_err();
-        assert!(matches!(err, LoadWorkspacePackagesError::InvalidPackageId(_)));
-    }
-
-    #[test]
-    #[serial]
-    fn load_workspace_packages_reports_not_staged_when_dir_missing() {
-        // Setup: pristine workspace root with no `target/streamlib-plugins/`
-        // staging. Helper must surface the actionable PackageNotStaged
-        // error rather than papering over the missing dir with a
-        // generic load_project failure.
-        let tmp = tempfile::tempdir().unwrap();
-        let key = "STREAMLIB_WORKSPACE_ROOT";
-        let prev = std::env::var_os(key);
-        unsafe {
-            std::env::set_var(key, tmp.path());
-        }
-        let runtime = Runner::new().expect("Runner::new");
-        let err = runtime
-            .load_workspace_packages(["@tatolab/camera"])
-            .unwrap_err();
-        unsafe {
-            match prev {
-                Some(v) => std::env::set_var(key, v),
-                None => std::env::remove_var(key),
-            }
-        }
-        assert!(matches!(
-            err,
-            LoadWorkspacePackagesError::PackageNotStaged { ref name, ref expected_path }
-                if name == "@tatolab/camera"
-                && expected_path.ends_with("tatolab__camera")
-        ));
-    }
-
-    #[test]
-    #[serial]
-    fn load_workspace_packages_returns_invalid_id_before_filesystem_probe() {
-        // Validation order: every id must parse BEFORE the helper
-        // calls `resolve_workspace_root`. Without the env-var override
-        // pointing at a bogus path that can't be a workspace, the
-        // unparseable id must still surface as `InvalidPackageId`
-        // (not the cargo-locate-project-derived `WorkspaceRootNotFound`).
-        // Mentally reverting the eager parse-loop to interleaved
-        // parse-then-resolve would flip the verdict to
-        // WorkspaceRootNotFound on this fixture, so the test catches
-        // the regression.
-        let tmp = tempfile::tempdir().unwrap();
-        let bogus = tmp.path().join("nowhere");
-        let key = "STREAMLIB_WORKSPACE_ROOT";
-        let prev = std::env::var_os(key);
-        unsafe {
-            std::env::set_var(key, &bogus);
-        }
-        let runtime = Runner::new().expect("Runner::new");
-        let err = runtime
-            .load_workspace_packages(["bad-no-at"])
-            .unwrap_err();
-        unsafe {
-            match prev {
-                Some(v) => std::env::set_var(key, v),
-                None => std::env::remove_var(key),
-            }
-        }
-        assert!(
-            matches!(
-                err,
-                LoadWorkspacePackagesError::InvalidPackageId(ref s) if s == "bad-no-at"
-            ),
-            "expected InvalidPackageId surfaced before workspace resolution, got: {err:?}"
-        );
-    }
 
     #[cfg(target_os = "linux")]
     #[test]
@@ -3192,665 +1140,7 @@ mod tests {
         );
     }
 
-    /// Path-style dep recursion: `runtime.load_project(A)` must walk into
-    /// `B` (declared as `path: ../b`) and parse its manifest. The negative
-    /// counterpart below proves recursion actually happens — not just that
-    /// the parser tolerates the structured shape.
-    #[test]
-    #[serial]
-    fn test_load_project_recurses_into_path_dep() {
-        let runtime = Runner::new().unwrap();
-        let tmp = tempfile::tempdir().unwrap();
 
-        // Project A — empty processors, declares path dep to ../b
-        let a = tmp.path().join("a");
-        std::fs::create_dir(&a).unwrap();
-        std::fs::write(
-            a.join("streamlib.yaml"),
-            r#"
-package:
-  org: tatolab
-  name: a
-  version: "0.1.0"
-dependencies:
-  "@tatolab/b":
-    path: ../b
-"#,
-        )
-        .unwrap();
-
-        // Project B — leaf, empty processors
-        let b = tmp.path().join("b");
-        std::fs::create_dir(&b).unwrap();
-        std::fs::write(
-            b.join("streamlib.yaml"),
-            r#"
-package:
-  org: tatolab
-  name: b
-  version: "0.1.0"
-"#,
-        )
-        .unwrap();
-
-        runtime
-            .load_project(&a)
-            .expect("load_project should recurse into path dep without error");
-    }
-
-    /// `Runner::load_project` reads each entry in `streamlib.yaml`'s
-    /// `schemas:` list and registers the YAML body with the engine's
-    /// runtime schema registry, so `get_embedded_schema_definition`
-    /// resolves the body and `max_payload_bytes_for_port_spec` returns
-    /// the value declared in `metadata.max_payload_bytes`.
-    /// Mentally reverting the `register_package_schemas(...)` call in
-    /// `load_project` would make this test fail because the registered
-    /// schema would be invisible to the lookup paths.
-    #[test]
-    #[serial]
-    fn test_load_project_registers_package_schemas_for_runtime_lookup() {
-        let runtime = Runner::new().unwrap();
-        let tmp = tempfile::tempdir().unwrap();
-
-        // Build a minimal package with a single schema. The package's
-        // `schemas:` list points at a schema file declaring
-        // `metadata.type` (new shape) + `metadata.max_payload_bytes`.
-        let pkg = tmp.path().join("pkg-with-schema");
-        std::fs::create_dir(&pkg).unwrap();
-        std::fs::create_dir(pkg.join("schemas")).unwrap();
-        std::fs::write(
-            pkg.join("schemas/my_test_config.yaml"),
-            "metadata:\n  type: MyTestConfig\n  max_payload_bytes: 8192\n",
-        )
-        .unwrap();
-        std::fs::write(
-            pkg.join("streamlib.yaml"),
-            r#"
-package:
-  org: tatolab
-  name: test-load-project-registers-schemas
-  version: "0.1.0"
-
-# Post-#767: `schemas:` is a name-keyed map of `Local { file }` /
-# `External { package }` declarations — replaces the legacy
-# `Vec<PathBuf>` shape.
-schemas:
-  MyTestConfig:
-    file: schemas/my_test_config.yaml
-"#,
-        )
-        .unwrap();
-
-        let canonical =
-            "@tatolab/test-load-project-registers-schemas/MyTestConfig";
-
-        // Pre-condition: schema not in registry.
-        assert!(
-            crate::core::embedded_schemas::get_embedded_schema_definition(canonical).is_none(),
-            "fresh canonical id must not exist before load_project"
-        );
-
-        runtime
-            .load_project(&pkg)
-            .expect("load_project must succeed for schemas-only package");
-
-        // Post-condition: lookup resolves both forms; payload bytes
-        // come from the schema's metadata, not the iceoryx2 default.
-        let body = crate::core::embedded_schemas::get_embedded_schema_definition(canonical)
-            .expect("registered schema must be discoverable post-load");
-        assert!(body.contains("MyTestConfig"));
-        let port_spec = streamlib_processor_schema::PortSchemaSpec::Specific(
-            streamlib_idents::SchemaIdent::new(
-                streamlib_idents::Org::new("tatolab").unwrap(),
-                streamlib_idents::Package::new(
-                    "test-load-project-registers-schemas",
-                )
-                .unwrap(),
-                streamlib_idents::TypeName::new("MyTestConfig").unwrap(),
-                streamlib_idents::SemVer::new(1, 0, 0),
-            ),
-        );
-        assert_eq!(
-            crate::core::embedded_schemas::max_payload_bytes_for_port_spec(&port_spec).unwrap(),
-            8192,
-            "max_payload_bytes_for_port_spec must read metadata declared by the loaded package"
-        );
-    }
-
-    /// Negative pair to the test above: when `B`'s manifest is missing,
-    /// the recursion must fail and propagate the error. Mentally reverting
-    /// the recursion in `load_project` would make this test pass falsely
-    /// — that's why both are needed.
-    #[test]
-    #[serial]
-    fn test_load_project_path_dep_missing_manifest_propagates_error() {
-        let runtime = Runner::new().unwrap();
-        let tmp = tempfile::tempdir().unwrap();
-
-        let a = tmp.path().join("a");
-        std::fs::create_dir(&a).unwrap();
-        std::fs::write(
-            a.join("streamlib.yaml"),
-            r#"
-package:
-  org: tatolab
-  name: a
-  version: "0.1.0"
-dependencies:
-  "@tatolab/missing":
-    path: ../does-not-exist
-"#,
-        )
-        .unwrap();
-
-        let result = runtime.load_project(&a);
-        assert!(
-            result.is_err(),
-            "load_project must error when a path dep target has no streamlib.yaml"
-        );
-    }
-
-    /// Consumer-scoped `patch:` resolution: when a consumer declares a
-    /// registry-form dep AND its own `patch:` block has an entry for that
-    /// dep pointing at a local path, `load_project` recurses into the
-    /// patched location. Mirrors Cargo's `[patch.crates-io]` semantics
-    /// but per-consumer (no workspace walk-up). Mentally reverting the
-    /// patch lookup in `resolve_dependency_path` makes this fail because
-    /// the registry arm would fall through to the installed-cache tier
-    /// (which is empty in this test).
-    #[test]
-    #[serial]
-    fn test_load_project_resolves_registry_dep_via_consumer_patch() {
-        let runtime = Runner::new().unwrap();
-        let tmp = tempfile::tempdir().unwrap();
-
-        // Patched dep target.
-        let b = tmp.path().join("b");
-        std::fs::create_dir(&b).unwrap();
-        std::fs::write(
-            b.join("streamlib.yaml"),
-            "package:\n  org: tatolab\n  name: b\n  version: \"0.1.0\"\n",
-        )
-        .unwrap();
-
-        // Consumer in a sibling dir, declares a registry-style dep AND a
-        // path-flavor patch in its own yaml. No tree-level state.
-        let a = tmp.path().join("a");
-        std::fs::create_dir(&a).unwrap();
-        std::fs::write(
-            a.join("streamlib.yaml"),
-            r#"
-package:
-  org: tatolab
-  name: a
-  version: "0.1.0"
-dependencies:
-  "@tatolab/b": "^0.1.0"
-patch:
-  "@tatolab/b":
-    path: ../b
-"#,
-        )
-        .unwrap();
-
-        runtime
-            .load_project(&a)
-            .expect("consumer-scoped patch must resolve the registry dep to ../b/");
-    }
-
-    /// Git-flavor patch resolution: when a consumer's `patch:` block
-    /// declares a `git:` entry pinned at a specific rev, `load_project`
-    /// clones the URL via the shared `streamlib_idents::fetch_git`
-    /// helper (same code the build-time resolver uses) and recurses
-    /// into the checkout. Mentally swapping the git arm in
-    /// `resolve_consumer_patch` for the previous "git/registry not
-    /// supported" error would make this test surface that error
-    /// instead of succeeding.
-    #[test]
-    #[serial]
-    fn test_load_project_resolves_git_patch_via_shared_helper() {
-        // Build a minimal local git repo with a `streamlib.yaml` that
-        // declares `@tatolab/b`. `git clone <local-path>` works without
-        // a network — this is a real fetch through the same helper the
-        // resolver uses, exercised end-to-end.
-        let tmp = tempfile::tempdir().unwrap();
-        let repo = tmp.path().join("b-repo");
-        std::fs::create_dir(&repo).unwrap();
-
-        let run_git = |args: &[&str]| {
-            let status = std::process::Command::new("git")
-                .args(args)
-                .current_dir(&repo)
-                .status()
-                .expect("git invocation");
-            assert!(status.success(), "git {:?} failed", args);
-        };
-
-        run_git(&["init", "--quiet"]);
-        run_git(&["config", "user.email", "test@example.com"]);
-        run_git(&["config", "user.name", "test"]);
-        std::fs::write(
-            repo.join("streamlib.yaml"),
-            "package:\n  org: tatolab\n  name: b\n  version: \"0.1.0\"\n",
-        )
-        .unwrap();
-        run_git(&["add", "streamlib.yaml"]);
-        run_git(&["commit", "--quiet", "-m", "initial"]);
-        let rev_output = std::process::Command::new("git")
-            .args(["rev-parse", "HEAD"])
-            .current_dir(&repo)
-            .output()
-            .expect("git rev-parse");
-        let rev = String::from_utf8(rev_output.stdout).unwrap().trim().to_string();
-
-        // Sandbox STREAMLIB_HOME so the git clone lands in tempdir, not
-        // the real user cache (and to avoid leaking checkouts across
-        // test runs).
-        let sandbox = tempfile::tempdir().unwrap();
-        let prev_home = std::env::var_os("STREAMLIB_HOME");
-        // SAFETY: `#[serial]` serializes every Runner test in this
-        // module, so concurrent env-var mutation can't tear other tests.
-        unsafe {
-            std::env::set_var("STREAMLIB_HOME", sandbox.path());
-        }
-        let _restore = StreamlibHomeRestore(prev_home);
-
-        // Consumer in a separate tempdir, declares a registry-style dep
-        // and a git-flavor patch pointing at the local repo.
-        let consumer = tempfile::tempdir().unwrap();
-        std::fs::write(
-            consumer.path().join("streamlib.yaml"),
-            format!(
-                r#"
-package:
-  org: tatolab
-  name: consumer
-  version: "0.1.0"
-dependencies:
-  "@tatolab/b": "^0.1.0"
-patch:
-  "@tatolab/b":
-    git: "{}"
-    rev: "{}"
-"#,
-                repo.display(),
-                rev,
-            ),
-        )
-        .unwrap();
-
-        let runtime = Runner::new().unwrap();
-        runtime
-            .load_project(consumer.path())
-            .expect("git patch must clone the local repo and recurse into it");
-    }
-
-    /// Strict patch validation: when the consumer's `patch:` block points
-    /// at a path that doesn't exist, `load_project` errors clearly so the
-    /// dev knows immediately to fix the manifest. npm/wrangler-style
-    /// strictness — the "make the right way easy and the wrong way hard"
-    /// rule from CLAUDE.md applied to manifest validation. Mentally
-    /// reverting the existence check in `resolve_consumer_patch` would
-    /// surface a downstream "manifest not found" error from `load_project`
-    /// instead of a clear "patch path doesn't exist" error.
-    #[test]
-    #[serial]
-    fn test_load_project_strict_errors_on_missing_patch_path() {
-        let runtime = Runner::new().unwrap();
-        let tmp = tempfile::tempdir().unwrap();
-        std::fs::write(
-            tmp.path().join("streamlib.yaml"),
-            r#"
-package:
-  org: tatolab
-  name: a
-  version: "0.1.0"
-dependencies:
-  "@tatolab/b": "^0.1.0"
-patch:
-  "@tatolab/b":
-    path: ./does-not-exist
-"#,
-        )
-        .unwrap();
-
-        let err = runtime
-            .load_project(tmp.path())
-            .expect_err("missing patch path must error strictly");
-        let msg = format!("{err}");
-        assert!(
-            msg.contains("@tatolab/b"),
-            "error must surface the canonical dep ref, got: {msg}"
-        );
-        assert!(
-            msg.contains("does-not-exist") && msg.contains("does not exist"),
-            "error must call out the missing patch path, got: {msg}"
-        );
-    }
-
-    /// Drops a previously-saved `STREAMLIB_HOME` environment variable
-    /// state when the test scope ends, so a sandboxed `STREAMLIB_HOME`
-    /// override doesn't leak into the next `#[serial]` test.
-    struct StreamlibHomeRestore(Option<std::ffi::OsString>);
-    impl Drop for StreamlibHomeRestore {
-        fn drop(&mut self) {
-            // SAFETY: `#[serial]` makes every test in this module
-            // exclusive — no concurrent reader of `STREAMLIB_HOME`.
-            unsafe {
-                match self.0.take() {
-                    Some(v) => std::env::set_var("STREAMLIB_HOME", v),
-                    None => std::env::remove_var("STREAMLIB_HOME"),
-                }
-            }
-        }
-    }
-
-    #[test]
-    #[serial]
-    fn test_load_project_resolves_registry_dep_via_installed_cache() {
-        // Sandbox: STREAMLIB_HOME → tempdir so the test doesn't interact
-        // with the real `~/.streamlib/packages.yaml`.
-        let sandbox = tempfile::tempdir().unwrap();
-        let prev_home = std::env::var_os("STREAMLIB_HOME");
-        // SAFETY: `#[serial]` serializes every Runner test in this
-        // module, so concurrent env-var mutation can't tear other tests.
-        unsafe {
-            std::env::set_var("STREAMLIB_HOME", sandbox.path());
-        }
-        let _restore = StreamlibHomeRestore(prev_home);
-
-        // The "extracted slpkg cache" directory the InstalledPackageManifest
-        // entry will point at. Contains the dep package's manifest.
-        let cache_root = sandbox.path().join("cache/packages");
-        std::fs::create_dir_all(&cache_root).unwrap();
-        let dep_cache_dir = cache_root.join("b-0.1.0");
-        std::fs::create_dir(&dep_cache_dir).unwrap();
-        std::fs::write(
-            dep_cache_dir.join("streamlib.yaml"),
-            "package:\n  org: tatolab\n  name: b\n  version: \"0.1.0\"\n",
-        )
-        .unwrap();
-
-        // Pre-populate the installed-package manifest with a canonical-key
-        // entry for `@tatolab/b` pointing at the cache dir above.
-        let mut installed = crate::core::config::InstalledPackageManifest::default();
-        installed.add(crate::core::config::InstalledPackageEntry {
-            name: streamlib_idents::PackageRef::new(
-                streamlib_processor_schema::Org::new("tatolab").unwrap(),
-                streamlib_processor_schema::Package::new("b").unwrap(),
-            ),
-            version: streamlib_processor_schema::SemVer::new(0, 1, 0),
-            description: None,
-            installed_from: "test".into(),
-            installed_at: "1970-01-01T00:00:00Z".into(),
-            cache_dir: "b-0.1.0".to_string(),
-        });
-        installed.save().unwrap();
-
-        // Customer-shape consumer: declares the dep canonically with NO
-        // `patch:` block — exactly the yaml shape that ships in slpkgs.
-        // Resolution must fall through to the installed-package cache.
-        let consumer = tempfile::tempdir().unwrap();
-        std::fs::write(
-            consumer.path().join("streamlib.yaml"),
-            r#"
-package:
-  org: tatolab
-  name: consumer
-  version: "0.1.0"
-dependencies:
-  "@tatolab/b": "^0.1.0"
-"#,
-        )
-        .unwrap();
-
-        let runtime = Runner::new().unwrap();
-        runtime
-            .load_project(consumer.path())
-            .expect("registry dep must resolve via installed-package cache");
-    }
-
-    /// When neither a workspace `[patch]` entry nor an installed-package
-    /// cache hit covers a registry dep, the runtime must surface an
-    /// actionable error that names the canonical key and points the user
-    /// at `streamlib pkg install`. The error is the runtime's last-resort
-    /// signal that resolution exhausted both tiers; mentally reverting
-    /// either tier (workspace lookup, installed-cache lookup) would make
-    /// this test surface a different error path, breaking the contract.
-    #[test]
-    #[serial]
-    fn test_load_project_unresolvable_registry_dep_errors_actionably() {
-        let runtime = Runner::new().unwrap();
-        let tmp = tempfile::tempdir().unwrap();
-
-        let a = tmp.path().join("a");
-        std::fs::create_dir(&a).unwrap();
-        std::fs::write(
-            a.join("streamlib.yaml"),
-            r#"
-package:
-  org: tatolab
-  name: a
-  version: "0.1.0"
-dependencies:
-  "@tatolab/missing": "^1.0.0"
-"#,
-        )
-        .unwrap();
-
-        let err = runtime
-            .load_project(&a)
-            .expect_err("unresolvable registry dep must error");
-        let msg = format!("{}", err);
-        assert!(
-            msg.contains("@tatolab/missing"),
-            "error must surface the canonical `@org/name` key, got: {msg}"
-        );
-        assert!(
-            msg.contains("streamlib pkg install") || msg.contains("workspace"),
-            "error must point at the resolution paths the user can act on, got: {msg}"
-        );
-    }
-
-    /// Plugin cdylib resolution: when a project declares Rust runtime
-    /// processors, `load_project` reads from `lib/<host_triple>/...`.
-    /// Mismatched-triple subdirs must produce a clear error naming both
-    /// the host triple and the triples actually present, so the user
-    /// can diagnose immediately ("this slpkg was packed for an
-    /// incompatible host, repack on a matching one"). Mentally
-    /// reverting the `triple_dir = lib_dir.join(host_triple)` line in
-    /// `load_project` would surface a non-existent host's flat-`lib/`
-    /// resolution and the test would not see the host triple in the
-    /// error message.
-    #[test]
-    #[serial]
-    fn test_load_project_rust_dylib_missing_host_triple_surfaces_available_triples() {
-        let runtime = Runner::new().unwrap();
-        let tmp = tempfile::tempdir().unwrap();
-
-        let pkg = tmp.path().join("pkg");
-        std::fs::create_dir(&pkg).unwrap();
-        std::fs::write(
-            pkg.join("streamlib.yaml"),
-            r#"
-package:
-  org: tatolab
-  name: triple-mismatch-pkg
-  version: "0.1.0"
-processors:
-  - name: TestProcessor
-    version: 1.0.0
-    description: "Test"
-    runtime: rust
-    execution: manual
-    inputs:
-      - name: video_in
-        schema: any
-    outputs:
-      - name: video_out
-        schema: any
-"#,
-        )
-        .unwrap();
-
-        // Plant a dylib under a deliberately-wrong triple so
-        // `list_available_triples` has something to surface and the
-        // error message can prove the triple-keyed resolution actually
-        // looked at the right subdir.
-        let wrong_triple = "wrong-arch-unknown-elsewhere-gnu";
-        let wrong_dir = pkg.join("lib").join(wrong_triple);
-        std::fs::create_dir_all(&wrong_dir).unwrap();
-        std::fs::write(wrong_dir.join("libfake.so"), b"not-a-real-dylib").unwrap();
-
-        let err = runtime
-            .load_project(&pkg)
-            .expect_err("missing host-triple subdir must error");
-        let msg = format!("{}", err);
-        assert!(
-            msg.contains(host_target_triple()),
-            "error must name the host triple so the user sees what was expected, got: {msg}"
-        );
-        assert!(
-            msg.contains(wrong_triple),
-            "error must list the triples that ARE present so the user sees what the slpkg was packed for, got: {msg}"
-        );
-    }
-
-    /// Positive pair to the triple-mismatch test above: when the
-    /// host-triple subdir exists and contains a file with the right
-    /// extension, `load_project` reaches the dlopen step. The dylib is
-    /// junk bytes (not a real cdylib), so dlopen fails — but the error
-    /// proves path resolution succeeded BEFORE dlopen. Mentally
-    /// reverting the triple lookup to flat-`lib/` would surface a
-    /// different error variant ("No .so file found in lib/") and this
-    /// test would not see the dlopen / STREAMLIB_PLUGIN-symbol message.
-    #[test]
-    #[serial]
-    fn test_load_project_rust_dylib_resolves_host_triple_then_dlopens() {
-        let runtime = Runner::new().unwrap();
-        let tmp = tempfile::tempdir().unwrap();
-
-        let pkg = tmp.path().join("pkg");
-        std::fs::create_dir(&pkg).unwrap();
-        std::fs::write(
-            pkg.join("streamlib.yaml"),
-            r#"
-package:
-  org: tatolab
-  name: triple-match-pkg
-  version: "0.1.0"
-processors:
-  - name: TestProcessor
-    version: 1.0.0
-    description: "Test"
-    runtime: rust
-    execution: manual
-    inputs:
-      - name: video_in
-        schema: any
-    outputs:
-      - name: video_out
-        schema: any
-"#,
-        )
-        .unwrap();
-
-        let triple_dir = pkg.join("lib").join(host_target_triple());
-        std::fs::create_dir_all(&triple_dir).unwrap();
-        let dylib_ext = if cfg!(target_os = "macos") {
-            "dylib"
-        } else if cfg!(target_os = "windows") {
-            "dll"
-        } else {
-            "so"
-        };
-        let dylib_path = triple_dir.join(format!("libfake.{}", dylib_ext));
-        std::fs::write(&dylib_path, b"not-a-real-dylib").unwrap();
-
-        let err = runtime
-            .load_project(&pkg)
-            .expect_err("junk dylib must fail at dlopen, not at path resolution");
-        let msg = format!("{}", err);
-        // Surface evidence that path resolution succeeded: the error
-        // must mention the dylib file itself (proving the loader got
-        // past `read_dir` and into `libloading::Library::new`). The
-        // earlier-shape error "No .so file found in lib/" would name
-        // the directory, not the file — so this assertion locks the
-        // resolution-then-dlopen ordering.
-        assert!(
-            msg.contains("libfake"),
-            "error must reference the dylib file (proving path resolution reached dlopen), got: {msg}"
-        );
-        assert!(
-            !msg.contains("No .so file found")
-                && !msg.contains("No .dylib file found")
-                && !msg.contains("No .dll file found"),
-            "error must NOT be the 'no dylib found' variant (path resolution succeeded), got: {msg}"
-        );
-    }
-
-    /// Schemas-only packages have no `lib/` directory at all and must
-    /// load cleanly — the loader's empty-processors short-circuit at
-    /// the top of `load_project` returns before touching the lib path.
-    /// Reverting that short-circuit would make this test surface a
-    /// "Failed to read lib/<triple>" error instead.
-    #[test]
-    #[serial]
-    fn test_load_project_schemas_only_skips_lib_lookup() {
-        let runtime = Runner::new().unwrap();
-        let tmp = tempfile::tempdir().unwrap();
-
-        let pkg = tmp.path().join("schemas-only");
-        std::fs::create_dir(&pkg).unwrap();
-        std::fs::write(
-            pkg.join("streamlib.yaml"),
-            r#"
-package:
-  org: tatolab
-  name: schemas-only-pkg
-  version: "0.1.0"
-"#,
-        )
-        .unwrap();
-
-        // Deliberately omit `lib/` — a schemas-only package has no
-        // platform-specific content, and the loader must NOT try to
-        // resolve a host-triple subdir.
-        runtime
-            .load_project(&pkg)
-            .expect("schemas-only package must load without touching lib/");
-    }
-
-    /// Unit lock for the `list_available_triples` helper:
-    /// returns directory-name strings sorted, ignores non-directory
-    /// entries, returns empty when `lib/` itself is missing. Reverting
-    /// any of the filters would either surface non-triple noise or
-    /// crash on missing-dir.
-    #[test]
-    fn list_available_triples_filters_to_subdirs_and_sorts() {
-        let tmp = tempfile::tempdir().unwrap();
-        let lib = tmp.path().join("lib");
-
-        // Missing lib/ → empty list, no error.
-        assert!(list_available_triples(&lib).unwrap().is_empty());
-
-        std::fs::create_dir(&lib).unwrap();
-        // Two triple-named subdirs + one non-dir file at the same level
-        // (a stray README or similar). The helper must list the two
-        // subdirs sorted, omit the file.
-        std::fs::create_dir(lib.join("aarch64-apple-darwin")).unwrap();
-        std::fs::create_dir(lib.join("x86_64-unknown-linux-gnu")).unwrap();
-        std::fs::write(lib.join("README.md"), b"stray").unwrap();
-
-        let triples = list_available_triples(&lib).unwrap();
-        assert_eq!(
-            triples,
-            vec![
-                "aarch64-apple-darwin".to_string(),
-                "x86_64-unknown-linux-gnu".to_string(),
-            ]
-        );
-    }
 
     #[test]
     #[serial]
@@ -4102,950 +1392,4 @@ package:
         }
     }
 
-    // =========================================================================
-    // add_module / remove_module — imperative module API
-    // =========================================================================
-
-    mod add_module_tests {
-        use super::*;
-        use streamlib_idents::{ModuleIdent, Org, Package, SemVer, SemVerRange};
-
-        /// RAII guard that restores `STREAMLIB_WORKSPACE_ROOT` and
-        /// `STREAMLIB_HOME` to their pre-test values on drop. Every
-        /// `add_module` test isolates both env vars to a tempdir to
-        /// keep the host's real workspace / installed cache out of the
-        /// test surface (per the side-effect-cleanup discipline in
-        /// `docs/testing.md` — flag tests that mutate global state).
-        struct AddModuleEnvGuard {
-            workspace_prev: Option<std::ffi::OsString>,
-            home_prev: Option<std::ffi::OsString>,
-        }
-
-        impl AddModuleEnvGuard {
-            fn install(workspace_root: &std::path::Path, home_root: &std::path::Path) -> Self {
-                let workspace_prev = std::env::var_os("STREAMLIB_WORKSPACE_ROOT");
-                let home_prev = std::env::var_os("STREAMLIB_HOME");
-                unsafe {
-                    std::env::set_var("STREAMLIB_WORKSPACE_ROOT", workspace_root);
-                    std::env::set_var("STREAMLIB_HOME", home_root);
-                }
-                Self { workspace_prev, home_prev }
-            }
-        }
-
-        impl Drop for AddModuleEnvGuard {
-            fn drop(&mut self) {
-                unsafe {
-                    match self.workspace_prev.take() {
-                        Some(v) => std::env::set_var("STREAMLIB_WORKSPACE_ROOT", v),
-                        None => std::env::remove_var("STREAMLIB_WORKSPACE_ROOT"),
-                    }
-                    match self.home_prev.take() {
-                        Some(v) => std::env::set_var("STREAMLIB_HOME", v),
-                        None => std::env::remove_var("STREAMLIB_HOME"),
-                    }
-                }
-            }
-        }
-
-        /// Build a schemas-only `streamlib.yaml` at `dir` with the
-        /// given org/name/version. Schemas-only avoids the dylib-
-        /// loading branch — `add_module` resolution + version-range
-        /// matching is what we want to lock here, not the cdylib
-        /// mechanics that `load_workspace_packages` already covers.
-        ///
-        /// When `schema` is provided, emits a `schemas:` block plus
-        /// the named schema YAML under `schemas/<type_snake>.yaml`
-        /// so post-load callers can verify the schema actually
-        /// registered (locks "`load_project` ran" — not just "resolver
-        /// + range check ran").
-        fn write_schemas_only_manifest(
-            dir: &std::path::Path,
-            org: &str,
-            name: &str,
-            version: &str,
-            schema: Option<&str>,
-        ) {
-            let body = if let Some(type_name) = schema {
-                let stem = type_name.to_ascii_lowercase();
-                std::fs::create_dir_all(dir.join("schemas")).unwrap();
-                std::fs::write(
-                    dir.join("schemas").join(format!("{stem}.yaml")),
-                    format!(
-                        "metadata:\n  type: {type_name}\n  max_payload_bytes: 4096\n"
-                    ),
-                )
-                .unwrap();
-                format!(
-                    "package:\n  org: {org}\n  name: {name}\n  version: \"{version}\"\n\
-                     schemas:\n  {type_name}:\n    file: schemas/{stem}.yaml\n"
-                )
-            } else {
-                format!(
-                    "package:\n  org: {org}\n  name: {name}\n  version: \"{version}\"\n"
-                )
-            };
-            std::fs::write(dir.join("streamlib.yaml"), body)
-                .expect("write streamlib.yaml");
-        }
-
-        /// Stage a package under
-        /// `<workspace>/target/streamlib-plugins/<org>__<name>/`
-        /// and return the staged dir. Mirrors what
-        /// `cargo xtask build-plugins` produces.
-        fn stage_workspace_package(
-            workspace_root: &std::path::Path,
-            org: &str,
-            name: &str,
-            version: &str,
-            schema: Option<&str>,
-        ) -> std::path::PathBuf {
-            let staged = workspace_root
-                .join("target")
-                .join("streamlib-plugins")
-                .join(format!("{org}__{name}"));
-            std::fs::create_dir_all(&staged).expect("mkdir staged");
-            write_schemas_only_manifest(&staged, org, name, version, schema);
-            staged
-        }
-
-        #[test]
-        #[serial]
-        fn add_module_resolves_workspace_stage_and_loads() {
-            // The schema name is unique to this test so the runtime
-            // schema registry (a process-wide static) can be probed
-            // after `add_module` to confirm load_project actually
-            // ran — mentally reverting the `self.load_project(...)`
-            // call inside `add_module` to `Ok(())` would leave the
-            // earlier (resolver-only) assertion green but make the
-            // post-load schema lookup fail.
-            const TYPE_NAME: &str = "AddModuleResolvesStageSchema";
-            let tmp = tempfile::tempdir().unwrap();
-            let home = tempfile::tempdir().unwrap();
-            let _staged = stage_workspace_package(
-                tmp.path(),
-                "tatolab",
-                "add-module-stage",
-                "1.2.3",
-                Some(TYPE_NAME),
-            );
-
-            let _guard = AddModuleEnvGuard::install(tmp.path(), home.path());
-            let runtime = Runner::new().expect("Runner::new");
-
-            // Pre-condition: schema is NOT in the registry yet.
-            let canonical = format!("@tatolab/add-module-stage/{TYPE_NAME}");
-            assert!(
-                crate::core::embedded_schemas::get_embedded_schema_definition(&canonical)
-                    .is_none(),
-                "schema must not exist before add_module"
-            );
-
-            runtime
-                .add_module(ModuleIdent::new(
-                    Org::new("tatolab").unwrap(),
-                    Package::new("add-module-stage").unwrap(),
-                    SemVerRange::Caret(SemVer::new(1, 0, 0)),
-                ))
-                .expect("workspace-staged add_module must succeed");
-
-            // Post-condition: schema IS registered. This is the
-            // "load_project actually ran" assertion the resolver-only
-            // tests didn't lock.
-            assert!(
-                crate::core::embedded_schemas::get_embedded_schema_definition(&canonical)
-                    .is_some(),
-                "schema must be registered after add_module — load_project did not run if this fails"
-            );
-        }
-
-        #[test]
-        #[serial]
-        fn add_module_any_version_succeeds_against_staged_v1() {
-            const TYPE_NAME: &str = "AddModuleAnyVersionSchema";
-            let tmp = tempfile::tempdir().unwrap();
-            let home = tempfile::tempdir().unwrap();
-            let _staged = stage_workspace_package(
-                tmp.path(),
-                "tatolab",
-                "add-module-any",
-                "0.4.0",
-                Some(TYPE_NAME),
-            );
-
-            let _guard = AddModuleEnvGuard::install(tmp.path(), home.path());
-            let runtime = Runner::new().expect("Runner::new");
-
-            let canonical = format!("@tatolab/add-module-any/{TYPE_NAME}");
-            assert!(
-                crate::core::embedded_schemas::get_embedded_schema_definition(&canonical)
-                    .is_none(),
-                "schema must not exist before add_module"
-            );
-
-            runtime
-                .add_module(ModuleIdent::any(
-                    Org::new("tatolab").unwrap(),
-                    Package::new("add-module-any").unwrap(),
-                ))
-                .expect("any-version add_module must succeed against staged 0.4.0");
-
-            assert!(
-                crate::core::embedded_schemas::get_embedded_schema_definition(&canonical)
-                    .is_some(),
-                "any-version add_module must invoke load_project; schema registry is the witness",
-            );
-        }
-
-        #[test]
-        #[serial]
-        fn add_module_rejects_version_range_mismatch() {
-            // Staged 1.0.0 but ident asks for ^2.0.0 — must surface
-            // VersionRangeUnsatisfied, NOT a generic "module not
-            // found" or a silently-succeeded load. Mentally reverting
-            // the range check would let the wrong version load.
-            let tmp = tempfile::tempdir().unwrap();
-            let home = tempfile::tempdir().unwrap();
-            let _staged = stage_workspace_package(
-                tmp.path(),
-                "tatolab",
-                "add-module-range",
-                "1.0.0",
-                None,
-            );
-
-            let _guard = AddModuleEnvGuard::install(tmp.path(), home.path());
-            let runtime = Runner::new().expect("Runner::new");
-            let err = runtime
-                .add_module(ModuleIdent::new(
-                    Org::new("tatolab").unwrap(),
-                    Package::new("add-module-range").unwrap(),
-                    SemVerRange::Caret(SemVer::new(2, 0, 0)),
-                ))
-                .expect_err("range mismatch must error");
-
-            assert!(
-                matches!(err, AddModuleError::VersionRangeUnsatisfied { found, .. } if found == SemVer::new(1, 0, 0)),
-                "expected VersionRangeUnsatisfied(1.0.0), got: {err:?}",
-            );
-        }
-
-        #[test]
-        #[serial]
-        fn add_module_rejects_identity_mismatch_on_staged_yaml() {
-            // Staged dir says `@vendor/other` but the requested ident
-            // is `@tatolab/add-module-identity`. The directory-name
-            // convention (`<org>__<name>/`) would point us there, but
-            // the manifest contents disagree. Surface
-            // ManifestIdentityMismatch rather than blindly loading.
-            let tmp = tempfile::tempdir().unwrap();
-            let home = tempfile::tempdir().unwrap();
-            let staged = tmp
-                .path()
-                .join("target")
-                .join("streamlib-plugins")
-                .join("tatolab__add-module-identity");
-            std::fs::create_dir_all(&staged).unwrap();
-            write_schemas_only_manifest(&staged, "vendor", "other", "1.0.0", None);
-
-            let _guard = AddModuleEnvGuard::install(tmp.path(), home.path());
-            let runtime = Runner::new().expect("Runner::new");
-            let err = runtime
-                .add_module(ModuleIdent::any(
-                    Org::new("tatolab").unwrap(),
-                    Package::new("add-module-identity").unwrap(),
-                ))
-                .expect_err("clobbered staged manifest must error");
-
-            assert!(
-                matches!(err, AddModuleError::ManifestIdentityMismatch { ref actual, .. } if actual == "@vendor/other"),
-                "expected ManifestIdentityMismatch(@vendor/other), got: {err:?}",
-            );
-        }
-
-        #[test]
-        #[serial]
-        fn add_module_reports_module_not_found_when_unstaged_and_uncached() {
-            // No workspace stage AND no installed cache entry. Must
-            // surface the actionable `ModuleNotFound` rather than a
-            // generic "load_project failed" or a panic. The
-            // `STREAMLIB_HOME` redirect isolates the test from the
-            // host's real installed cache — without it, a stale entry
-            // for the canary name would mask the failure.
-            let tmp = tempfile::tempdir().unwrap();
-            let home = tempfile::tempdir().unwrap();
-            let _guard = AddModuleEnvGuard::install(tmp.path(), home.path());
-
-            let runtime = Runner::new().expect("Runner::new");
-            let err = runtime
-                .add_module(ModuleIdent::any(
-                    Org::new("tatolab").unwrap(),
-                    Package::new("add-module-missing").unwrap(),
-                ))
-                .expect_err("missing module must error");
-
-            assert!(
-                matches!(err, AddModuleError::ModuleNotFound { ref package } if package.org.as_str() == "tatolab" && package.name.as_str() == "add-module-missing"),
-                "expected ModuleNotFound(@tatolab/add-module-missing), got: {err:?}",
-            );
-        }
-
-        #[test]
-        #[serial]
-        fn add_module_surfaces_workspace_root_typo() {
-            // STREAMLIB_WORKSPACE_ROOT pointing at a non-existent path
-            // is a user typo — surface it as WorkspaceRootInvalid
-            // rather than silently falling through to the installed
-            // cache (which would either find an unrelated entry, or
-            // mis-report ModuleNotFound when the real problem is the
-            // env var). Mirrors load_workspace_packages's behavior on
-            // the same condition.
-            //
-            // Install the RAII env guard against a tempdir first (its
-            // Drop will restore both env vars on exit / panic), then
-            // override STREAMLIB_WORKSPACE_ROOT to the bogus path
-            // inside the guard's lifetime. If `add_module` or
-            // `expect_err` panics, the guard still restores the
-            // original values — no env leak into sibling tests.
-            let home = tempfile::tempdir().unwrap();
-            let placeholder_workspace = tempfile::tempdir().unwrap();
-            let _guard =
-                AddModuleEnvGuard::install(placeholder_workspace.path(), home.path());
-            unsafe {
-                std::env::set_var(
-                    "STREAMLIB_WORKSPACE_ROOT",
-                    "/nonexistent/path/that/does/not/exist",
-                );
-            }
-
-            let runtime = Runner::new().expect("Runner::new");
-            let err = runtime
-                .add_module(ModuleIdent::any(
-                    Org::new("tatolab").unwrap(),
-                    Package::new("add-module-typo-canary").unwrap(),
-                ))
-                .expect_err("typo'd env var must error");
-
-            assert!(
-                matches!(err, AddModuleError::WorkspaceRootInvalid { ref env_value } if env_value == "/nonexistent/path/that/does/not/exist"),
-                "expected WorkspaceRootInvalid, got: {err:?}",
-            );
-        }
-
-        // =====================================================================
-        // ModuleResolverStrategy conformance — one test per variant.
-        //
-        // Each test exercises `add_module_with(ident, strategy)` against a
-        // fixture set up to satisfy exactly that strategy's resolver tier,
-        // then asserts the module loads. Mentally reverting the strategy
-        // resolver's hit-arm for the variant under test would surface the
-        // typed error variant the negative sibling pins.
-        // =====================================================================
-
-        #[test]
-        #[serial]
-        fn add_module_with_default_chain_falls_through_workspace_to_installed_cache() {
-            // Default chain semantics: try workspace stage first, then
-            // installed cache. With NO workspace stage but a populated
-            // installed cache, the strategy must hit tier 2.
-            const TYPE_NAME: &str = "DefaultChainFallsThroughSchema";
-            let workspace = tempfile::tempdir().unwrap();
-            let home = tempfile::tempdir().unwrap();
-            let _guard = AddModuleEnvGuard::install(workspace.path(), home.path());
-
-            // Populate the installed cache with @tatolab/dc-fallthrough@0.1.0.
-            let cache_root = home.path().join("cache/packages");
-            std::fs::create_dir_all(&cache_root).unwrap();
-            let dep_cache_dir = cache_root.join("dc-fallthrough-0.1.0");
-            std::fs::create_dir(&dep_cache_dir).unwrap();
-            write_schemas_only_manifest(
-                &dep_cache_dir,
-                "tatolab",
-                "dc-fallthrough",
-                "0.1.0",
-                Some(TYPE_NAME),
-            );
-            let mut installed =
-                crate::core::config::InstalledPackageManifest::default();
-            installed.add(crate::core::config::InstalledPackageEntry {
-                name: streamlib_idents::PackageRef::new(
-                    Org::new("tatolab").unwrap(),
-                    Package::new("dc-fallthrough").unwrap(),
-                ),
-                version: SemVer::new(0, 1, 0),
-                description: None,
-                installed_from: "test".into(),
-                installed_at: "1970-01-01T00:00:00Z".into(),
-                cache_dir: "dc-fallthrough-0.1.0".to_string(),
-            });
-            installed.save().unwrap();
-
-            let runtime = Runner::new().expect("Runner::new");
-            let canonical = format!("@tatolab/dc-fallthrough/{TYPE_NAME}");
-            assert!(
-                crate::core::embedded_schemas::get_embedded_schema_definition(&canonical)
-                    .is_none()
-            );
-
-            runtime
-                .add_module_with(
-                    ModuleIdent::any(
-                        Org::new("tatolab").unwrap(),
-                        Package::new("dc-fallthrough").unwrap(),
-                    ),
-                    ModuleResolverStrategy::DefaultChain,
-                )
-                .expect("DefaultChain must reach installed cache when workspace has no stage");
-
-            assert!(
-                crate::core::embedded_schemas::get_embedded_schema_definition(&canonical)
-                    .is_some(),
-                "DefaultChain's installed-cache tier did not actually run load",
-            );
-        }
-
-        #[test]
-        #[serial]
-        fn add_module_with_workspace_staged_strategy_hits_stage_dir_only() {
-            const TYPE_NAME: &str = "WorkspaceStagedOnlySchema";
-            let workspace = tempfile::tempdir().unwrap();
-            let home = tempfile::tempdir().unwrap();
-            let _staged = stage_workspace_package(
-                workspace.path(),
-                "tatolab",
-                "ws-only",
-                "1.0.0",
-                Some(TYPE_NAME),
-            );
-            let _guard = AddModuleEnvGuard::install(workspace.path(), home.path());
-            let runtime = Runner::new().expect("Runner::new");
-            let canonical = format!("@tatolab/ws-only/{TYPE_NAME}");
-            assert!(
-                crate::core::embedded_schemas::get_embedded_schema_definition(&canonical)
-                    .is_none()
-            );
-            runtime
-                .add_module_with(
-                    ModuleIdent::any(
-                        Org::new("tatolab").unwrap(),
-                        Package::new("ws-only").unwrap(),
-                    ),
-                    ModuleResolverStrategy::WorkspaceStaged,
-                )
-                .expect("WorkspaceStaged must hit the stage dir");
-            assert!(
-                crate::core::embedded_schemas::get_embedded_schema_definition(&canonical)
-                    .is_some()
-            );
-        }
-
-        #[test]
-        #[serial]
-        fn add_module_with_workspace_staged_strategy_surfaces_miss() {
-            // No workspace stage and no fallback: WorkspaceStaged must
-            // surface the typed `WorkspaceStageMiss` rather than silently
-            // falling through. Mentally reverting the
-            // [`WorkspaceStageMiss`] return in the resolver would make
-            // this surface a generic `ModuleNotFound`.
-            let workspace = tempfile::tempdir().unwrap();
-            let home = tempfile::tempdir().unwrap();
-            let _guard = AddModuleEnvGuard::install(workspace.path(), home.path());
-            let runtime = Runner::new().expect("Runner::new");
-            let err = runtime
-                .add_module_with(
-                    ModuleIdent::any(
-                        Org::new("tatolab").unwrap(),
-                        Package::new("ws-only-miss").unwrap(),
-                    ),
-                    ModuleResolverStrategy::WorkspaceStaged,
-                )
-                .expect_err("missing stage dir must error");
-            assert!(
-                matches!(
-                    err,
-                    AddModuleError::WorkspaceStageMiss { ref package, ref expected_path }
-                        if package.name.as_str() == "ws-only-miss"
-                            && expected_path.ends_with("tatolab__ws-only-miss")
-                ),
-                "expected WorkspaceStageMiss, got: {err:?}",
-            );
-        }
-
-        #[test]
-        #[serial]
-        fn add_module_with_installed_cache_strategy_hits_cache_only() {
-            const TYPE_NAME: &str = "InstalledCacheOnlySchema";
-            // Set workspace to something OTHER than where a stage would
-            // be — strategy is InstalledCache so workspace must not be
-            // consulted. (We still need an existing dir for the env
-            // guard since WorkspaceRootInvalid would otherwise fire.)
-            let workspace = tempfile::tempdir().unwrap();
-            let home = tempfile::tempdir().unwrap();
-            let cache_root = home.path().join("cache/packages");
-            std::fs::create_dir_all(&cache_root).unwrap();
-            let dep_cache_dir = cache_root.join("ic-only-1.0.0");
-            std::fs::create_dir(&dep_cache_dir).unwrap();
-            write_schemas_only_manifest(
-                &dep_cache_dir,
-                "tatolab",
-                "ic-only",
-                "1.0.0",
-                Some(TYPE_NAME),
-            );
-            let mut installed =
-                crate::core::config::InstalledPackageManifest::default();
-            installed.add(crate::core::config::InstalledPackageEntry {
-                name: streamlib_idents::PackageRef::new(
-                    Org::new("tatolab").unwrap(),
-                    Package::new("ic-only").unwrap(),
-                ),
-                version: SemVer::new(1, 0, 0),
-                description: None,
-                installed_from: "test".into(),
-                installed_at: "1970-01-01T00:00:00Z".into(),
-                cache_dir: "ic-only-1.0.0".to_string(),
-            });
-            let _guard = AddModuleEnvGuard::install(workspace.path(), home.path());
-            installed.save().unwrap();
-
-            let runtime = Runner::new().expect("Runner::new");
-            let canonical = format!("@tatolab/ic-only/{TYPE_NAME}");
-            runtime
-                .add_module_with(
-                    ModuleIdent::any(
-                        Org::new("tatolab").unwrap(),
-                        Package::new("ic-only").unwrap(),
-                    ),
-                    ModuleResolverStrategy::InstalledCache,
-                )
-                .expect("InstalledCache strategy must hit the cache");
-            assert!(
-                crate::core::embedded_schemas::get_embedded_schema_definition(&canonical)
-                    .is_some()
-            );
-        }
-
-        #[test]
-        #[serial]
-        fn add_module_with_manifest_directory_loads_arbitrary_dir() {
-            // ManifestDirectory bypasses ident-keyed resolution
-            // entirely — the path IS the source of truth. Used by
-            // load_project() and by per-call patch overrides.
-            const TYPE_NAME: &str = "ManifestDirectorySchema";
-            let workspace = tempfile::tempdir().unwrap();
-            let home = tempfile::tempdir().unwrap();
-            let arbitrary = tempfile::tempdir().unwrap();
-            write_schemas_only_manifest(
-                arbitrary.path(),
-                "tatolab",
-                "md-arbitrary",
-                "0.7.2",
-                Some(TYPE_NAME),
-            );
-            let _guard = AddModuleEnvGuard::install(workspace.path(), home.path());
-            let runtime = Runner::new().expect("Runner::new");
-            let canonical = format!("@tatolab/md-arbitrary/{TYPE_NAME}");
-            runtime
-                .add_module_with(
-                    ModuleIdent::any(
-                        Org::new("tatolab").unwrap(),
-                        Package::new("md-arbitrary").unwrap(),
-                    ),
-                    ModuleResolverStrategy::ManifestDirectory {
-                        path: arbitrary.path().to_path_buf(),
-                    },
-                )
-                .expect("ManifestDirectory must load the arbitrary dir");
-            assert!(
-                crate::core::embedded_schemas::get_embedded_schema_definition(&canonical)
-                    .is_some()
-            );
-        }
-
-        #[test]
-        #[serial]
-        fn add_module_with_manifest_directory_surfaces_missing_dir() {
-            let workspace = tempfile::tempdir().unwrap();
-            let home = tempfile::tempdir().unwrap();
-            let _guard = AddModuleEnvGuard::install(workspace.path(), home.path());
-            let runtime = Runner::new().expect("Runner::new");
-            let err = runtime
-                .add_module_with(
-                    ModuleIdent::any(
-                        Org::new("tatolab").unwrap(),
-                        Package::new("md-missing").unwrap(),
-                    ),
-                    ModuleResolverStrategy::ManifestDirectory {
-                        path: std::path::PathBuf::from("/nonexistent/does-not-exist"),
-                    },
-                )
-                .expect_err("missing manifest dir must error");
-            assert!(
-                matches!(err, AddModuleError::ManifestDirectoryMissing { .. }),
-                "expected ManifestDirectoryMissing, got: {err:?}",
-            );
-        }
-
-        #[test]
-        #[serial]
-        fn add_module_with_explicit_strategy_overrides_default_chain() {
-            // The per-call override IS the in-code equivalent of the
-            // manifest's `patch: path:` table — proves that an explicit
-            // strategy reaches the resolver and short-circuits the
-            // workspace-stage / installed-cache default chain even when
-            // BOTH default tiers would also have hit.
-            const TYPE_NAME: &str = "PerCallOverrideSchema";
-            let workspace = tempfile::tempdir().unwrap();
-            let home = tempfile::tempdir().unwrap();
-
-            // Stage a "wrong" version (0.1.0) under the workspace.
-            let _staged = stage_workspace_package(
-                workspace.path(),
-                "tatolab",
-                "per-call-override",
-                "0.1.0",
-                None,
-            );
-
-            // Also write the "right" version (9.0.0) at an arbitrary
-            // directory the override will point at, AND emit a schema
-            // unique to this version so the post-load registry probe
-            // can witness WHICH manifest actually loaded.
-            let overridden = tempfile::tempdir().unwrap();
-            write_schemas_only_manifest(
-                overridden.path(),
-                "tatolab",
-                "per-call-override",
-                "9.0.0",
-                Some(TYPE_NAME),
-            );
-
-            let _guard = AddModuleEnvGuard::install(workspace.path(), home.path());
-            let runtime = Runner::new().expect("Runner::new");
-
-            runtime
-                .add_module_with(
-                    ModuleIdent::any(
-                        Org::new("tatolab").unwrap(),
-                        Package::new("per-call-override").unwrap(),
-                    ),
-                    ModuleResolverStrategy::ManifestDirectory {
-                        path: overridden.path().to_path_buf(),
-                    },
-                )
-                .expect("explicit override must reach the resolver");
-
-            let canonical = format!("@tatolab/per-call-override/{TYPE_NAME}");
-            assert!(
-                crate::core::embedded_schemas::get_embedded_schema_definition(&canonical)
-                    .is_some(),
-                "explicit strategy didn't actually divert the resolver — schema from \
-                 overridden path is missing",
-            );
-        }
-
-        // =====================================================================
-        // Recursive dep walker — each declared dep routes through
-        // `add_module_with`, which means the new shape's contract is
-        // "single primitive, multiple entry points." Tests below
-        // exercise multi-dep walk, version-range mismatch surfaced from
-        // a dep, and the cycle-detection error.
-        // =====================================================================
-
-        #[test]
-        #[serial]
-        fn add_module_recurses_through_add_module_with_for_each_dep() {
-            // A → B → C — all three must register.
-            const TYPE_A: &str = "DepWalkAllThreeASchema";
-            const TYPE_B: &str = "DepWalkAllThreeBSchema";
-            const TYPE_C: &str = "DepWalkAllThreeCSchema";
-            let workspace = tempfile::tempdir().unwrap();
-            let home = tempfile::tempdir().unwrap();
-            let pkg_root = tempfile::tempdir().unwrap();
-
-            // Build three manifest dirs side-by-side; A path-deps B
-            // which path-deps C.
-            let a = pkg_root.path().join("a");
-            let b = pkg_root.path().join("b");
-            let c = pkg_root.path().join("c");
-            std::fs::create_dir(&a).unwrap();
-            std::fs::create_dir(&b).unwrap();
-            std::fs::create_dir(&c).unwrap();
-            std::fs::create_dir(a.join("schemas")).unwrap();
-            std::fs::create_dir(b.join("schemas")).unwrap();
-            std::fs::create_dir(c.join("schemas")).unwrap();
-            std::fs::write(
-                a.join("schemas/depwalkallthreeaschema.yaml"),
-                format!("metadata:\n  type: {TYPE_A}\n  max_payload_bytes: 4096\n"),
-            )
-            .unwrap();
-            std::fs::write(
-                b.join("schemas/depwalkallthreebschema.yaml"),
-                format!("metadata:\n  type: {TYPE_B}\n  max_payload_bytes: 4096\n"),
-            )
-            .unwrap();
-            std::fs::write(
-                c.join("schemas/depwalkallthreecschema.yaml"),
-                format!("metadata:\n  type: {TYPE_C}\n  max_payload_bytes: 4096\n"),
-            )
-            .unwrap();
-            std::fs::write(
-                a.join("streamlib.yaml"),
-                format!(
-                    "package:\n  org: tatolab\n  name: dwa-a\n  version: \"0.1.0\"\n\
-                     dependencies:\n  \"@tatolab/dwa-b\":\n    path: ../b\n\
-                     schemas:\n  {TYPE_A}:\n    file: schemas/depwalkallthreeaschema.yaml\n"
-                ),
-            )
-            .unwrap();
-            std::fs::write(
-                b.join("streamlib.yaml"),
-                format!(
-                    "package:\n  org: tatolab\n  name: dwa-b\n  version: \"0.1.0\"\n\
-                     dependencies:\n  \"@tatolab/dwa-c\":\n    path: ../c\n\
-                     schemas:\n  {TYPE_B}:\n    file: schemas/depwalkallthreebschema.yaml\n"
-                ),
-            )
-            .unwrap();
-            std::fs::write(
-                c.join("streamlib.yaml"),
-                format!(
-                    "package:\n  org: tatolab\n  name: dwa-c\n  version: \"0.1.0\"\n\
-                     schemas:\n  {TYPE_C}:\n    file: schemas/depwalkallthreecschema.yaml\n"
-                ),
-            )
-            .unwrap();
-
-            let _guard = AddModuleEnvGuard::install(workspace.path(), home.path());
-            let runtime = Runner::new().expect("Runner::new");
-            runtime
-                .add_module_with(
-                    ModuleIdent::any(
-                        Org::new("tatolab").unwrap(),
-                        Package::new("dwa-a").unwrap(),
-                    ),
-                    ModuleResolverStrategy::ManifestDirectory { path: a.clone() },
-                )
-                .expect("dep walker must succeed end-to-end");
-            for (pkg, ty) in [
-                ("dwa-a", TYPE_A),
-                ("dwa-b", TYPE_B),
-                ("dwa-c", TYPE_C),
-            ] {
-                let canonical = format!("@tatolab/{pkg}/{ty}");
-                assert!(
-                    crate::core::embedded_schemas::get_embedded_schema_definition(&canonical)
-                        .is_some(),
-                    "expected schema {canonical} registered after dep walk",
-                );
-            }
-        }
-
-        #[test]
-        #[serial]
-        fn add_module_detects_self_referential_cycle() {
-            // A depends on A (path: .). The recursive walker must surface
-            // `DependencyCycleDetected` rather than infinite-loop or
-            // panic on stack overflow. The dep-walker propagates the
-            // typed error unwrapped via bare `?`, so any
-            // `LoadProjectFailed`-shaped wrap would be a regression.
-            let workspace = tempfile::tempdir().unwrap();
-            let home = tempfile::tempdir().unwrap();
-            let pkg = tempfile::tempdir().unwrap();
-            std::fs::write(
-                pkg.path().join("streamlib.yaml"),
-                "package:\n  org: tatolab\n  name: cycle-self\n  version: \"0.1.0\"\n\
-                 dependencies:\n  \"@tatolab/cycle-self\":\n    path: .\n",
-            )
-            .unwrap();
-            let _guard = AddModuleEnvGuard::install(workspace.path(), home.path());
-            let runtime = Runner::new().expect("Runner::new");
-            let err = runtime
-                .add_module_with(
-                    ModuleIdent::any(
-                        Org::new("tatolab").unwrap(),
-                        Package::new("cycle-self").unwrap(),
-                    ),
-                    ModuleResolverStrategy::ManifestDirectory {
-                        path: pkg.path().to_path_buf(),
-                    },
-                )
-                .expect_err("self-referential cycle must error");
-            // Strict match: cycle detection surfaces unwrapped. The
-            // recursion-path Vec must contain the repeated vertex
-            // (`cycle-self`) as both the first and last entry so a
-            // reader sees `cycle-self → cycle-self`.
-            match err {
-                AddModuleError::DependencyCycleDetected { cycle } => {
-                    assert!(
-                        cycle.len() >= 2
-                            && cycle.first().unwrap().name.as_str() == "cycle-self"
-                            && cycle.last().unwrap().name.as_str() == "cycle-self",
-                        "expected cycle starting and ending at cycle-self, got: {cycle:?}",
-                    );
-                }
-                other => panic!("expected DependencyCycleDetected, got: {other:?}"),
-            }
-        }
-
-        #[test]
-        #[serial]
-        fn add_module_detects_mutual_cycle_a_to_b_to_a() {
-            // A path-deps B, B path-deps A. The recursion-path Vec
-            // must record the full A → B → A edge so the diagnostic
-            // shows exactly which transitive load re-entered.
-            let workspace = tempfile::tempdir().unwrap();
-            let home = tempfile::tempdir().unwrap();
-            let pkg_root = tempfile::tempdir().unwrap();
-            let a = pkg_root.path().join("a");
-            let b = pkg_root.path().join("b");
-            std::fs::create_dir(&a).unwrap();
-            std::fs::create_dir(&b).unwrap();
-            std::fs::write(
-                a.join("streamlib.yaml"),
-                "package:\n  org: tatolab\n  name: cycle-a\n  version: \"0.1.0\"\n\
-                 dependencies:\n  \"@tatolab/cycle-b\":\n    path: ../b\n",
-            )
-            .unwrap();
-            std::fs::write(
-                b.join("streamlib.yaml"),
-                "package:\n  org: tatolab\n  name: cycle-b\n  version: \"0.1.0\"\n\
-                 dependencies:\n  \"@tatolab/cycle-a\":\n    path: ../a\n",
-            )
-            .unwrap();
-            let _guard = AddModuleEnvGuard::install(workspace.path(), home.path());
-            let runtime = Runner::new().expect("Runner::new");
-            let err = runtime
-                .add_module_with(
-                    ModuleIdent::any(
-                        Org::new("tatolab").unwrap(),
-                        Package::new("cycle-a").unwrap(),
-                    ),
-                    ModuleResolverStrategy::ManifestDirectory { path: a.clone() },
-                )
-                .expect_err("mutual cycle must error");
-            match err {
-                AddModuleError::DependencyCycleDetected { cycle } => {
-                    let names: Vec<&str> = cycle.iter().map(|p| p.name.as_str()).collect();
-                    // First entry is the root we asked to load
-                    // (`cycle-a`); somewhere later in the path the
-                    // dep walker re-enters `cycle-a` after visiting
-                    // `cycle-b`. The repeated vertex is the last
-                    // entry, so `names.last()` must equal `cycle-a`
-                    // and `cycle-b` must appear between them.
-                    assert_eq!(names.first(), Some(&"cycle-a"));
-                    assert_eq!(names.last(), Some(&"cycle-a"));
-                    assert!(
-                        names.contains(&"cycle-b"),
-                        "expected cycle path to traverse cycle-b, got: {names:?}",
-                    );
-                }
-                other => panic!("expected DependencyCycleDetected, got: {other:?}"),
-            }
-        }
-
-        #[test]
-        #[serial]
-        fn add_module_surfaces_version_mismatch_propagating_from_dep() {
-            // A path-deps B; A's manifest pins B to "^2.0.0" but B's
-            // path target ships "1.0.0". For path-style declarations
-            // the declared range is `any` (path is authoritative), so a
-            // pure path dep does NOT trigger version mismatch.
-            // Instead, use a registry-style declaration patched with a
-            // path: the patch swaps the SOURCE while keeping the
-            // declared range — exactly Cargo's `[patch]` semantics —
-            // and the resulting version range check fires.
-            let workspace = tempfile::tempdir().unwrap();
-            let home = tempfile::tempdir().unwrap();
-            let pkg_root = tempfile::tempdir().unwrap();
-            let a = pkg_root.path().join("a");
-            let b = pkg_root.path().join("b");
-            std::fs::create_dir(&a).unwrap();
-            std::fs::create_dir(&b).unwrap();
-            std::fs::write(
-                a.join("streamlib.yaml"),
-                "package:\n  org: tatolab\n  name: vmm-a\n  version: \"0.1.0\"\n\
-                 dependencies:\n  \"@tatolab/vmm-b\": \"^2.0.0\"\n\
-                 patch:\n  \"@tatolab/vmm-b\":\n    path: ../b\n",
-            )
-            .unwrap();
-            std::fs::write(
-                b.join("streamlib.yaml"),
-                "package:\n  org: tatolab\n  name: vmm-b\n  version: \"1.0.0\"\n",
-            )
-            .unwrap();
-            let _guard = AddModuleEnvGuard::install(workspace.path(), home.path());
-            let runtime = Runner::new().expect("Runner::new");
-            let err = runtime
-                .add_module_with(
-                    ModuleIdent::any(
-                        Org::new("tatolab").unwrap(),
-                        Package::new("vmm-a").unwrap(),
-                    ),
-                    ModuleResolverStrategy::ManifestDirectory { path: a.clone() },
-                )
-                .expect_err("version mismatch from dep must propagate");
-            assert!(
-                matches!(err, AddModuleError::VersionRangeUnsatisfied { ref module, found, .. }
-                    if module.name.as_str() == "vmm-b" && found == SemVer::new(1, 0, 0)),
-                "expected VersionRangeUnsatisfied for vmm-b@1.0.0, got: {err:?}",
-            );
-        }
-
-        #[test]
-        #[serial]
-        fn add_module_default_path_keeps_default_chain_when_no_override() {
-            // Lock the negative half of the override contract: bare
-            // `add_module(ident)` keeps the default-chain behavior —
-            // proves the override path is the ONLY way to escape the
-            // default chain, and that the absence of an override
-            // doesn't silently morph into a different strategy.
-            const TYPE_NAME: &str = "DefaultChainStaysSchema";
-            let workspace = tempfile::tempdir().unwrap();
-            let home = tempfile::tempdir().unwrap();
-            let _staged = stage_workspace_package(
-                workspace.path(),
-                "tatolab",
-                "default-chain-stays",
-                "1.0.0",
-                Some(TYPE_NAME),
-            );
-            let _guard = AddModuleEnvGuard::install(workspace.path(), home.path());
-            let runtime = Runner::new().expect("Runner::new");
-            runtime
-                .add_module(ModuleIdent::any(
-                    Org::new("tatolab").unwrap(),
-                    Package::new("default-chain-stays").unwrap(),
-                ))
-                .expect("bare add_module must keep DefaultChain behavior");
-            let canonical = format!("@tatolab/default-chain-stays/{TYPE_NAME}");
-            assert!(
-                crate::core::embedded_schemas::get_embedded_schema_definition(&canonical)
-                    .is_some()
-            );
-        }
-
-        #[test]
-        fn remove_module_returns_hot_reload_lifecycle_deferral() {
-            // The stub returns a typed error rather than silently
-            // succeeding or panicking. AI agents reading `add_module`
-            // will reach for `remove_module`; the typed error points
-            // them at the milestone boundary.
-            let runtime = Runner::new().expect("Runner::new");
-            let err = runtime
-                .remove_module(ModuleIdent::any(
-                    Org::new("tatolab").unwrap(),
-                    Package::new("remove-module-stub").unwrap(),
-                ))
-                .expect_err("remove_module must error until hot-reload ships");
-            assert!(
-                matches!(
-                    err,
-                    RemoveModuleError::HotReloadLifecycleNotYetImplemented { ref module }
-                        if module.name.as_str() == "remove-module-stub"
-                ),
-                "got: {err:?}",
-            );
-        }
-    }
 }
